[worker] mcp split 4-6/6: per-tool modules extracted; mod.rs is wiring
Completes the src/mcp/mod.rs split (todo #105; stages 0-3 landed earlier):
the five remaining #[tool] methods move to per-tool modules with their own
#[tool_router] impls, merged into mod.rs's merged_tool_router; the internal
dispatch targets (find_definition, find_usages, file_outline, find_imports,
find_dependents, similar_chunks, literal_search, the semantic machinery and
status internals) move with them (pub(crate) where called cross-module).
mod.rs keeps the struct, ctors, routing/federation plumbing, rest handlers
and the merged router: 5619 -> ~2020 lines (from 8221 at the start of #105).

The stdout detector now scans every src/mcp source file so new per-tool
modules cannot escape it, and the registration pin proves all six tools
still route.

Mechanical refactor: zero behavior change. Full suite green (649 lib + 645
bins), clippy clean.
