+++ src/mcp/mod.rs
@@ -42,7 +42,7 @@ use rmcp::{
     handler::server::router::tool::ToolRouter,
     handler::server::wrapper::Parameters,
     model::{CallToolResult, Content, Implementation, ServerCapabilities, ServerInfo},
-    tool, tool_handler, tool_router, ErrorData as McpError, ServerHandler,
+    tool_handler, tool_router, ErrorData as McpError, ServerHandler,
 };
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
@@ -51,13 +51,20 @@ use std::sync::{Arc, Mutex};
 // Re-export types
 pub use types::*;
 
+mod explore;
 mod federation_helpers;
+mod find;
 mod find_impact;
+mod get_chunk;
+mod graph;
 mod helpers;
 mod instructions;
+mod literal_search;
 mod proxy;
 mod responses;
 mod runtime;
+mod search;
+mod status;
 
 // Re-export the extracted modules' items so every existing path keeps working:
 // `super::X` from sibling test files and `crate::mcp::X` from serve/cli.
@@ -1735,3616 +1742,6 @@ impl CodesearchService {
         let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
         CallToolResult::success(vec![Content::text(json)])
     }
-
-    // ─────────────────────────────────────────────────────────────────
-    // Consolidated tools (the primary 5-tool surface)
-    // ─────────────────────────────────────────────────────────────────
-
-    /// Unified search tool — dispatches to semantic or literal search based on `mode`.
-    #[tool(
-        description = "Unified code search. Set `mode` to choose the backend:\n\n- `semantic` (default): vector embeddings + BM25 FTS + exact-identifier boosting, fused with RRF. Best for conceptual queries, identifier lookups, and mixed natural-language + symbol queries.\n- `literal`: pure FTS, no embeddings. Fast and works without an embedding model. Sub-mode selection:\n  * Queries with operators, brackets, or punctuation (`foo = null`, `Vec<T>`, `return x;`, `a::b`) -> set `regex=true` and write the query as a regex. BM25 tokenizes on punctuation otherwise, producing noisy results.\n  * Multi-word exact phrases -> set `phrase=true`.\n  * Plain identifier lookups (`CodesearchService`) -> leave both false.\n\nFor semantic mode, optionally set `semantic_mode`: \"auto\" (default) | \"semantic\" | \"lexical\" | \"hybrid\".\nReturns metadata only by default (`compact=true`). Use `get_chunk` to read full code. Prefer `search(mode=\"literal\", regex=true)` over external grep/ripgrep for code patterns.\n\nIMPORTANT (multi-repo): always specify either `project` (single repo) or `group` (cross-repo). Omitting both in multi-repo mode returns a `scope_required` error with the list of available projects and groups. If the user has not indicated which repository to search, ask them to choose."
-    )]
-    async fn search(
-        &self,
-        Parameters(request): Parameters<SearchRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        tracing::info!(
-            "📥 search(query={:?}, mode={:?}, project={:?}, group={:?})",
-            request.query,
-            request.mode,
-            request.project,
-            request.group,
-        );
-
-        // Federation: when the query targets a group that resolves to one or more
-        // remote peers, merge local + remote results (RRF-interleave) instead of
-        // searching local repos only. Only `group` federates; `project` stays
-        // local because project aliases are instance-local.
-        if let Some(group) = request.group.as_deref() {
-            let cfg = self.federation_config();
-            if Self::group_has_remotes(&cfg, group) {
-                let remote_projects = cfg.group_remote_projects(group);
-                return self.federated_search(&request, &cfg, remote_projects).await;
-            }
-        }
-
-        // Project-level federation (mounted remote project): a `project` of the
-        // form "<peer>/<alias>" transparently routes to that single peer's own
-        // `<alias>` project — a 1-to-1 passthrough, as if the index were local.
-        // Local repos ALWAYS win a name clash: only route remotely when the name
-        // does not resolve to a local project.
-        if let Some(proj) = request.project.as_deref() {
-            let cfg = self.federation_config();
-            if cfg.resolve(proj).is_none() {
-                if let Some(crate::db_discovery::repos::Target::RemoteProject {
-                    peer_name,
-                    peer,
-                    remote_alias,
-                }) = cfg.resolve_remote_project(proj)
-                {
-                    return self
-                        .federated_project_search(&request, peer_name, peer, remote_alias)
-                        .await;
-                }
-            }
-        }
-
-        let mode = request.mode.as_deref().unwrap_or("semantic").to_lowercase();
-        match mode.as_str() {
-            "semantic" => {
-                // Delegate to the existing semantic_search implementation
-                let semantic_req = SemanticSearchRequest {
-                    query: request.query,
-                    limit: request.limit,
-                    compact: request.compact,
-                    filter_path: request.filter_path,
-                    mode: request.semantic_mode,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.semantic_search(Parameters(semantic_req)).await
-            }
-            "literal" => {
-                // Delegate to the existing literal_search implementation
-                let literal_req = LiteralSearchRequest {
-                    query: request.query,
-                    regex: request.regex,
-                    phrase: request.phrase,
-                    limit: request.limit,
-                    file_glob: request.file_glob,
-                    language: request.language,
-                    format: request.format,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.literal_search(Parameters(literal_req)).await
-            }
-            _ => Ok(CallToolResult::success(vec![Content::text(format!(
-                "Unknown search mode '{}'. Use `semantic` or `literal`.",
-                mode
-            ))])),
-        }
-    }
-
-    /// Unified symbol navigation — dispatches based on `kind`.
-    #[tool(
-        description = "Unified symbol navigation. Set `kind` to choose the action:\n\n- `definition` (default): locate where a symbol is defined (function, class, struct, etc.)\n- `usages`: find call-sites of a symbol via LEXICAL TEXT matching — hits may be docs/comments rather than code references; ranking puts source files first and a `note` field flags the precise upgrade path when one exists. On C#/TypeScript projects ALWAYS prefer `find_impact` for usages — it returns exact SCIP references, while this kind is only a text fallback\n- `imports`: list all imports/dependencies declared in a file (set `symbol` to the file path)\n- `dependents`: find all files that import or depend on a module, file, or symbol\n\nFor `imports`, set `symbol` to a file path. For other kinds, `symbol` is the symbol name.\n\nIMPORTANT (multi-repo): always specify either `project` (single repo) or `group` (cross-repo). Omitting both in multi-repo mode returns a `scope_required` error with the list of available projects and groups. If the user has not indicated which repository to search, ask them to choose."
-    )]
-    async fn find(
-        &self,
-        Parameters(request): Parameters<FindRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        let kind = request
-            .kind
-            .as_deref()
-            .unwrap_or("definition")
-            .to_lowercase();
-        tracing::info!(
-            "📥 find(symbol={:?}, kind={}, project={:?}, group={:?})",
-            request.symbol,
-            kind,
-            request.project,
-            request.group,
-        );
-        match kind.as_str() {
-            "definition" => {
-                let def_req = FindDefinitionRequest {
-                    symbol: request.symbol,
-                    kind: request.definition_kind,
-                    limit: request.limit,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.find_definition(Parameters(def_req)).await
-            }
-            "usages" => {
-                let usages_req = FindUsagesRequest {
-                    symbol: request.symbol,
-                    limit: request.limit,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.find_usages(Parameters(usages_req)).await
-            }
-            "imports" => {
-                let imports_req = FindImportsRequest {
-                    path: request.symbol,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.find_imports(Parameters(imports_req)).await
-            }
-            "dependents" => {
-                let dep_req = FindDependentsRequest {
-                    symbol_or_path: request.symbol,
-                    limit: request.limit,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.find_dependents(Parameters(dep_req)).await
-            }
-            _ => Ok(CallToolResult::success(vec![Content::text(format!(
-                "Unknown find kind '{}'. Use `definition`, `usages`, `imports`, or `dependents`.",
-                kind
-            ))])),
-        }
-    }
-
-    /// Unified exploration tool — dispatches based on `kind`.
-    #[tool(
-        description = "Unified code exploration. Set `kind` to choose the action:\n\n- `outline` (default): list all indexed top-level symbols in a file — kind, signature, and line range. Set `target` to a file path.\n- `similar`: find chunks semantically similar to a given chunk by its ID. Set `target` to the chunk_id (as string).\n\nIMPORTANT (multi-repo): always specify either `project` (single repo) or `group` (cross-repo). Omitting both in multi-repo mode returns a `scope_required` error with the list of available projects and groups. If the user has not indicated which repository to search, ask them to choose."
-    )]
-    async fn explore(
-        &self,
-        Parameters(request): Parameters<ExploreRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        let kind = request.kind.as_deref().unwrap_or("outline").to_lowercase();
-        tracing::info!(
-            "📥 explore(target={:?}, kind={}, project={:?})",
-            request.target,
-            kind,
-            request.project,
-        );
-        match kind.as_str() {
-            "outline" => {
-                let outline_req = FileOutlineRequest {
-                    path: request.target,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.file_outline(Parameters(outline_req)).await
-            }
-            "similar" => {
-                let chunk_id = match request.target.parse::<u32>() {
-                    Ok(id) => id,
-                    Err(_) => {
-                        return Ok(CallToolResult::success(vec![Content::text(format!(
-                            "For similar mode, `target` must be a numeric chunk_id, got: '{}'",
-                            request.target
-                        ))]));
-                    }
-                };
-                let similar_req = SimilarChunksRequest {
-                    chunk_id,
-                    limit: request.limit,
-                    project: request.project,
-                    group: request.group,
-                };
-                self.similar_chunks(Parameters(similar_req)).await
-            }
-            _ => Ok(CallToolResult::success(vec![Content::text(format!(
-                "Unknown explore kind '{}'. Use `outline` or `similar`.",
-                kind
-            ))])),
-        }
-    }
-
-    /// Unified status tool — dispatches based on `kind`.
-    #[tool(
-        description = "Unified status/info tool. Set `kind` to choose the action:\n\n- `index` (default): get the status of the local search index (model info, chunk count, readiness)\n- `projects`: list all registered projects/repositories, groups, and their index status"
-    )]
-    async fn status(
-        &self,
-        Parameters(request): Parameters<StatusRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        let kind = request.kind.as_deref().unwrap_or("index").to_lowercase();
-        tracing::info!("📥 status(kind={})", kind);
-        match kind.as_str() {
-            "index" => self.index_status_impl(request.project, request.group).await,
-            "projects" => self.list_projects().await,
-            _ => Ok(CallToolResult::success(vec![Content::text(format!(
-                "Unknown status kind '{}'. Use `index` or `projects`.",
-                kind
-            ))])),
-        }
-    }
-
-    // ─────────────────────────────────────────────────────────────────
-    // Internal implementations (called by consolidated tools above)
-    // ─────────────────────────────────────────────────────────────────
-
-    /// Internal: semantic/hybrid search implementation used by `search(mode="semantic")`.
-    async fn semantic_search(
-        &self,
-        Parameters(request): Parameters<SemanticSearchRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        // Resolve project/group routing (multi-store for group fan-out)
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "search")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        let limit = request.limit.unwrap_or(10);
-        let compact = request.compact.unwrap_or(true);
-        let mode = request.mode.as_deref().unwrap_or("auto");
-        let identifiers = detect_identifiers(&request.query);
-        let has_identifiers = !identifiers.is_empty();
-
-        tracing::debug!(
-            "MCP semantic_search: query='{}', limit={}, compact={}, mode='{}', multi={}",
-            request.query,
-            limit,
-            compact,
-            mode,
-            ctx.is_multi
-        );
-
-        // Ensure database exists (skip if serve-mode with routed stores)
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        // === Multi-store group fan-out ===
-        if ctx.is_multi {
-            return self
-                .semantic_search_multi(
-                    &request,
-                    &identifiers,
-                    limit,
-                    compact,
-                    ctx.stores_vec.unwrap(),
-                    ctx.store_aliases.as_ref().unwrap(),
-                    &ctx.alias_roots,
-                )
-                .await;
-        }
-
-        // === Mode: "lexical" — FTS only, no embedding ===
-        if mode == "lexical" {
-            tracing::debug!("MCP: mode=lexical — skipping embedding service");
-            return self
-                .semantic_search_lexical(
-                    &request,
-                    &identifiers,
-                    limit,
-                    compact,
-                    ctx.stores,
-                    ctx.project_alias.as_deref(),
-                    &ctx.alias_roots,
-                )
-                .await;
-        }
-
-        // === Modes: "semantic", "hybrid", "auto" — require embedding ===
-        let query_embedding = {
-            let mut service_guard = match self.get_embedding_service() {
-                Ok(g) => g,
-                Err(e) => {
-                    tracing::error!("MCP: Failed to get embedding service: {:?}", e);
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error initializing embedding service: {e:#}"
-                    ))]));
-                }
-            };
-
-            let service = service_guard.as_mut().unwrap();
-            tracing::debug!("MCP: Embedding query...");
-            match service.embed_query(&request.query) {
-                Ok(e) => e,
-                Err(e) => {
-                    tracing::error!("MCP: Failed to embed query: {:?}", e);
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error embedding query: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        // Failures on this single-store path. The group fan-out has carried a
-        // warnings channel since the read-only incident; without the same thing
-        // here, `project=<alias>` — the form an agent uses most — still reports
-        // a broken store as an ordinary empty result.
-        let mut single_warnings: Vec<String> = Vec::new();
-
-        // Search vector store
-        let vector_results = match self
-            .with_vector_store_read_for(
-                |store| {
-                    store
-                        .search(&query_embedding, limit * 5)
-                        .context("Error searching vector store")
-                },
-                ctx.stores.clone(),
-            )
-            .await
-        {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::error!("MCP: Search failed: {:?}", e);
-                // Only "semantic" has no second backend to fall back on. In
-                // hybrid/auto the FTS half can still answer, so hard-failing
-                // here would throw away good results — the same mistake this
-                // branch already fixed once in the group fan-out.
-                //
-                // `{:#}` renders the whole anyhow chain. With plain `{}` the
-                // caller only ever saw the outermost `.context(...)` wrapper
-                // ("Error reading from project-routed vector store"), which
-                // hides the actual fault and makes remote diagnosis guesswork.
-                if mode == "semantic" {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error searching vector store: {:#}",
-                        e
-                    ))]));
-                }
-                single_warnings.push(format!("vector search failed: {e:#}"));
-                Vec::new()
-            }
-        };
-
-        tracing::debug!("MCP: Found {} vector results", vector_results.len());
-
-        // === Mode: "semantic" — vector only, skip FTS fusion ===
-        if mode == "semantic" {
-            tracing::debug!("MCP: mode=semantic — using vector results only");
-            let fused = vector_only(&vector_results);
-
-            let chunk_to_result: std::collections::HashMap<u32, &crate::vectordb::SearchResult> =
-                vector_results.iter().map(|r| (r.id, r)).collect();
-
-            let mut results: Vec<crate::vectordb::SearchResult> = Vec::new();
-            for f in fused.into_iter().take(limit) {
-                if let Some(result) = chunk_to_result.get(&f.chunk_id) {
-                    let mut r = (*result).clone();
-                    r.score = f.rrf_score;
-                    results.push(r);
-                }
-            }
-            return self.build_semantic_response(
-                results,
-                &request,
-                compact,
-                has_identifiers,
-                ctx.project_alias.as_deref(),
-                &ctx.alias_roots,
-                &single_warnings,
-            );
-        }
-
-        // === Modes: "hybrid" | "auto" — full hybrid search ===
-        let structural_intent = detect_structural_intent(&request.query);
-        let (vector_k, fts_k) = adapt_rrf_k(&request.query);
-
-        tracing::debug!(
-            "MCP: Query analysis - identifiers: {:?}, structural_intent: {:?}, rrf_k: ({}, {})",
-            identifiers,
-            structural_intent,
-            vector_k,
-            fts_k
-        );
-
-        // Perform FTS search and fusion
-        let mut results = match self
-            .with_fts_store_read_for(
-                |fts_store| {
-                    let fts_results = fts_store
-                        .search(&request.query, limit * 5, structural_intent)
-                        .context("Error searching FTS store")?;
-
-                    let fused = if identifiers.is_empty() {
-                        rrf_fusion(&vector_results, &fts_results, vector_k as f32)
-                    } else {
-                        let mut all_exact: Vec<crate::fts::FtsResult> = Vec::new();
-                        for ident in &identifiers {
-                            if let Ok(exact) =
-                                fts_store.search_exact(ident, limit * 3, structural_intent)
-                            {
-                                for r in exact {
-                                    if !all_exact.iter().any(|e| e.chunk_id == r.chunk_id) {
-                                        all_exact.push(r);
-                                    }
-                                }
-                            }
-                        }
-
-                        tracing::debug!(
-                            "MCP: FTS found {} results, exact found {} results",
-                            fts_results.len(),
-                            all_exact.len()
-                        );
-
-                        rrf_fusion_with_exact(
-                            &vector_results,
-                            &fts_results,
-                            &all_exact,
-                            vector_k as f32,
-                            fts_k as f32,
-                            EXACT_MATCH_RRF_K,
-                        )
-                    };
-
-                    Ok(fused)
-                },
-                ctx.stores.clone(),
-            )
-            .await
-        {
-            Ok(fused) => {
-                // Map FusedResult back to SearchResult
-                let chunk_to_result: std::collections::HashMap<
-                    u32,
-                    &crate::vectordb::SearchResult,
-                > = vector_results.iter().map(|r| (r.id, r)).collect();
-
-                let mut mapped: Vec<crate::vectordb::SearchResult> = Vec::new();
-                for f in fused.into_iter().take(limit) {
-                    if let Some(result) = chunk_to_result.get(&f.chunk_id) {
-                        let mut r = (*result).clone();
-                        r.score = f.rrf_score;
-                        mapped.push(r);
-                    }
-                }
-                mapped
-            }
-            Err(e) => {
-                tracing::warn!("MCP: FTS store unavailable, using vector-only: {:?}", e);
-                // Degrading to vector-only is correct, but it must be VISIBLE:
-                // a caller that gets half a hybrid search with no signal cannot
-                // tell it from a complete one.
-                single_warnings.push(format!("lexical (FTS) search failed: {e:#}"));
-                vector_results.into_iter().take(limit).collect()
-            }
-        };
-
-        // Apply language boost
-        if let Some((_, _, Some(primary_lang))) = crate::search::read_metadata(&self.db_path) {
-            for result in &mut results {
-                let file_lang = format!(
-                    "{:?}",
-                    Language::from_path(std::path::Path::new(&result.path))
-                );
-                if file_lang.to_lowercase() == primary_lang.to_lowercase() {
-                    result.score *= 1.2;
-                }
-            }
-            results.sort_by(|a, b| {
-                b.score
-                    .partial_cmp(&a.score)
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            });
-        }
-
-        // Apply kind boost
-        if let Some(target_kind) = structural_intent {
-            boost_kind(&mut results, target_kind);
-        }
-
-        // Auto-fallback: if hybrid search returned very few results for a code-like query,
-        // run literal FTS and merge missing chunks.
-        if results.len() < 3 && has_identifiers {
-            tracing::debug!(
-                "Auto-fallback: semantic returned {} results, trying literal",
-                results.len()
-            );
-
-            let literal_results = self
-                .with_fts_store_read_for(
-                    |fts_store| fts_store.search(&request.query, limit, None),
-                    ctx.stores.clone(),
-                )
-                .await
-                .unwrap_or_default();
-
-            let mut existing_ids: std::collections::HashSet<u32> =
-                results.iter().map(|r| r.id).collect();
-
-            for fts in literal_results {
-                if results.len() >= limit {
-                    break;
-                }
-                if existing_ids.contains(&fts.chunk_id) {
-                    continue;
-                }
-
-                let maybe_resolved = match self
-                    .with_vector_store_read_for(
-                        |store| {
-                            // `Ok(None)` means "this store does not hold that
-                            // chunk" — a normal miss to skip. `Err` means the
-                            // store is broken and must propagate: flattening
-                            // the two silently dropped every remaining literal
-                            // hit whenever the vector store was down, turning a
-                            // dead store into an ordinary-looking short result.
-                            let chunk = match store.get_chunk(fts.chunk_id)? {
-                                Some(c) => c,
-                                None => return Ok(None),
-                            };
-                            Ok(Some(crate::vectordb::SearchResult {
-                                id: fts.chunk_id,
-                                content: chunk.content,
-                                path: chunk.path,
-                                start_line: chunk.start_line,
-                                end_line: chunk.end_line,
-                                kind: chunk.kind,
-                                signature: chunk.signature,
-                                docstring: chunk.docstring,
-                                context: chunk.context,
-                                hash: chunk.hash,
-                                distance: 0.0,
-                                score: fts.score,
-                                context_prev: chunk.context_prev,
-                                context_next: chunk.context_next,
-                            }))
-                        },
-                        ctx.stores.clone(),
-                    )
-                    .await
-                {
-                    Ok(resolved) => resolved,
-                    Err(e) => {
-                        // The old `.ok()` here folded a dead store into "no
-                        // more literal hits" with zero signal to the caller —
-                        // the exact false negative `single_warnings` exists
-                        // for. Note it and stop: every further lookup against
-                        // this store would fail the same way.
-                        single_warnings.push(format!("literal-hit chunk lookup failed: {e:#}"));
-                        break;
-                    }
-                };
-
-                if let Some(resolved) = maybe_resolved {
-                    existing_ids.insert(resolved.id);
-                    results.push(resolved);
-                }
-            }
-        }
-
-        tracing::debug!("MCP: Final {} results after hybrid search", results.len());
-        self.build_semantic_response(
-            results,
-            &request,
-            compact,
-            has_identifiers,
-            ctx.project_alias.as_deref(),
-            &ctx.alias_roots,
-            &single_warnings,
-        )
-    }
-
-    // === Helper methods (not exposed as tools) ===
-
-    /// Multi-store semantic search: fan out across all stores, merge raw vector/FTS
-    /// results, then apply RRF fusion.
-    #[allow(clippy::too_many_arguments)]
-    async fn semantic_search_multi(
-        &self,
-        request: &SemanticSearchRequest,
-        identifiers: &[String],
-        limit: usize,
-        compact: bool,
-        stores: Vec<Arc<SharedStores>>,
-        aliases: &[String],
-        alias_roots: &std::collections::HashMap<String, String>,
-    ) -> Result<CallToolResult, McpError> {
-        let mode = request.mode.as_deref().unwrap_or("auto");
-        let structural_intent = detect_structural_intent(&request.query);
-
-        // === Lexical mode: FTS only across all stores ===
-        if mode == "lexical" {
-            // Lexical has no second backend, so a failed store here is invisible
-            // unless it is reported: the query simply looks like it found nothing.
-            let mut lexical_warnings: Vec<String> = Vec::new();
-
-            let outcome = self
-                .with_fts_store_read_multi(
-                    |fts_store| fts_store.search(&request.query, limit * 5, structural_intent),
-                    stores.clone(),
-                    aliases,
-                )
-                .await
-                .unwrap_or_default();
-            if !outcome.failures.is_empty() {
-                tracing::error!(
-                    "MCP: lexical fan-out degraded — {} of {} repo(s) failed: {:?}",
-                    outcome.failures.len(),
-                    stores.len(),
-                    outcome.failures
-                );
-                lexical_warnings.extend(outcome.warnings("literal search"));
-            }
-            let fts_results = outcome.results;
-
-            // Also do exact search if identifiers detected
-            let mut all_fts = fts_results;
-            for ident in identifiers {
-                let exact_outcome = self
-                    .with_fts_store_read_multi(
-                        |fts_store| fts_store.search_exact(ident, limit * 3, structural_intent),
-                        stores.clone(),
-                        aliases,
-                    )
-                    .await
-                    .unwrap_or_default();
-                lexical_warnings.extend(exact_outcome.warnings("exact-identifier search"));
-                merge_exact_into_fts(&mut all_fts, exact_outcome.results);
-            }
-
-            all_fts.sort_by(|a, b| {
-                b.score
-                    .partial_cmp(&a.score)
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            });
-
-            let results = self
-                .resolve_fts_to_search_results_multi(
-                    &all_fts,
-                    limit,
-                    &stores,
-                    aliases,
-                    &mut lexical_warnings,
-                )
-                .await;
-
-            if let Some(target_kind) = structural_intent {
-                // We need mutable results but we have them as vectordb::SearchResult
-                let mut mutable_results = results;
-                boost_kind(&mut mutable_results, target_kind);
-                return self.build_semantic_response(
-                    mutable_results,
-                    request,
-                    compact,
-                    !identifiers.is_empty(),
-                    None,
-                    alias_roots,
-                    &lexical_warnings,
-                );
-            }
-
-            return self.build_semantic_response(
-                results,
-                request,
-                compact,
-                !identifiers.is_empty(),
-                None,
-                alias_roots,
-                &lexical_warnings,
-            );
-        }
-
-        // === Modes requiring embedding: "semantic", "hybrid", "auto" ===
-        let query_embedding = {
-            let mut service_guard = match self.get_embedding_service() {
-                Ok(g) => g,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error initializing embedding service: {e:#}"
-                    ))]));
-                }
-            };
-            let service = service_guard.as_mut().unwrap();
-            match service.embed_query(&request.query) {
-                Ok(e) => e,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error embedding query: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        // Search vector stores across all repos
-        let outcome = self
-            .with_vector_store_read_multi(
-                |store| {
-                    store
-                        .search(&query_embedding, limit * 5)
-                        .context("Error searching vector store")
-                },
-                stores.clone(),
-                aliases,
-            )
-            .await;
-
-        // Warnings raised by the fan-out, carried into the response so the
-        // calling agent can tell "not in the corpus" from "that repo is down".
-        let mut search_warnings: Vec<String> = Vec::new();
-
-        let vector_results =
-            match outcome {
-                Ok(o) => {
-                    if !o.failures.is_empty() {
-                        tracing::error!(
-                            "MCP: vector fan-out degraded — {} of {} repo(s) failed: {:?}",
-                            o.failures.len(),
-                            stores.len(),
-                            o.failures
-                        );
-                        // Only "semantic" has no second backend to fall back on. In
-                        // hybrid/auto/lexical the FTS half can still answer, so
-                        // hard-failing here would throw away good results — the same
-                        // reason one broken repo does not abort the whole fan-out.
-                        if mode == "semantic" && o.results.is_empty() {
-                            let detail = o
-                                .failures
-                                .iter()
-                                .map(|(alias, err)| format!("  - {alias}: {err}"))
-                                .collect::<Vec<_>>()
-                                .join("\n");
-                            return Ok(CallToolResult::success(vec![Content::text(format!(
-                                "Error searching vector store: {} of {} repo(s) in scope failed \
-                             and none returned results:\n{}",
-                                o.failures.len(),
-                                stores.len(),
-                                detail
-                            ))]));
-                        }
-                        search_warnings.extend(o.failures.iter().map(|(alias, err)| {
-                            format!("repo '{alias}' vector search failed: {err}")
-                        }));
-                    }
-                    o.results
-                }
-                Err(e) => {
-                    tracing::error!("MCP: vector fan-out failed: {:?}", e);
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error searching vector store: {e:#}"
-                    ))]));
-                }
-            };
-
-        // === Mode: "semantic" — vector only ===
-        if mode == "semantic" {
-            let fused = vector_only(&vector_results);
-            let chunk_to_result: std::collections::HashMap<u32, &crate::vectordb::SearchResult> =
-                vector_results.iter().map(|r| (r.id, r)).collect();
-
-            let mut results: Vec<crate::vectordb::SearchResult> = Vec::new();
-            for f in fused.into_iter().take(limit) {
-                if let Some(result) = chunk_to_result.get(&f.chunk_id) {
-                    let mut r = (*result).clone();
-                    r.score = f.rrf_score;
-                    results.push(r);
-                }
-            }
-            return self.build_semantic_response(
-                results,
-                request,
-                compact,
-                !identifiers.is_empty(),
-                None,
-                alias_roots,
-                &search_warnings,
-            );
-        }
-
-        // === Modes: "hybrid" | "auto" — full hybrid search ===
-        let (vector_k, fts_k) = adapt_rrf_k(&request.query);
-
-        // FTS search across all stores. Its failures matter as much as the
-        // vector half's: during the cloud read-only incident literal search
-        // also returned 0 results for every affected vendor, and looked clean.
-        let fts_outcome = self
-            .with_fts_store_read_multi(
-                |fts_store| fts_store.search(&request.query, limit * 5, structural_intent),
-                stores.clone(),
-                aliases,
-            )
-            .await
-            .unwrap_or_default();
-        if !fts_outcome.failures.is_empty() {
-            tracing::error!(
-                "MCP: FTS fan-out degraded — {} of {} repo(s) failed: {:?}",
-                fts_outcome.failures.len(),
-                stores.len(),
-                fts_outcome.failures
-            );
-            search_warnings.extend(fts_outcome.warnings("literal search"));
-        }
-        let fts_results = fts_outcome.results;
-
-        // Exact identifier search across all stores
-        let all_exact = if !identifiers.is_empty() {
-            let mut exact_results: Vec<crate::fts::FtsResult> = Vec::new();
-            for ident in identifiers {
-                let exact_outcome = self
-                    .with_fts_store_read_multi(
-                        |fts_store| fts_store.search_exact(ident, limit * 3, structural_intent),
-                        stores.clone(),
-                        aliases,
-                    )
-                    .await
-                    .unwrap_or_default();
-                search_warnings.extend(exact_outcome.warnings("exact-identifier search"));
-                for r in exact_outcome.results {
-                    if !exact_results.iter().any(|e| e.chunk_id == r.chunk_id) {
-                        exact_results.push(r);
-                    }
-                }
-            }
-            exact_results
-        } else {
-            Vec::new()
-        };
-
-        // RRF fusion
-        let fused = if identifiers.is_empty() {
-            rrf_fusion(&vector_results, &fts_results, vector_k as f32)
-        } else {
-            rrf_fusion_with_exact(
-                &vector_results,
-                &fts_results,
-                &all_exact,
-                vector_k as f32,
-                fts_k as f32,
-                EXACT_MATCH_RRF_K,
-            )
-        };
-
-        // Map FusedResult back to SearchResult via chunk lookup across all stores
-        let chunk_to_result: std::collections::HashMap<u32, &crate::vectordb::SearchResult> =
-            vector_results.iter().map(|r| (r.id, r)).collect();
-
-        let mut mapped: Vec<crate::vectordb::SearchResult> = Vec::new();
-        for f in fused.into_iter().take(limit) {
-            if let Some(result) = chunk_to_result.get(&f.chunk_id) {
-                let mut r = (*result).clone();
-                r.score = f.rrf_score;
-                mapped.push(r);
-            } else {
-                // Chunk from FTS but not in vector results — resolve from stores
-                if let Some(resolved) = self
-                    .resolve_chunk_from_stores(
-                        f.chunk_id,
-                        f.rrf_score,
-                        &stores,
-                        aliases,
-                        &mut search_warnings,
-                    )
-                    .await
-                {
-                    mapped.push(resolved);
-                }
-            }
-        }
-
-        // Apply kind boost
-        if let Some(target_kind) = structural_intent {
-            boost_kind(&mut mapped, target_kind);
-        }
-
-        self.build_semantic_response(
-            mapped,
-            request,
-            compact,
-            !identifiers.is_empty(),
-            None,
-            alias_roots,
-            &search_warnings,
-        )
-    }
-
-    /// Resolve a single chunk from multiple stores (used for FTS-only hits in multi-store fusion).
-    async fn resolve_chunk_from_stores(
-        &self,
-        chunk_id: u32,
-        score: f32,
-        stores: &[Arc<SharedStores>],
-        aliases: &[String],
-        warnings: &mut Vec<String>,
-    ) -> Option<crate::vectordb::SearchResult> {
-        for (idx, store_arc) in stores.iter().enumerate() {
-            let store = store_arc.vector_store.read().await;
-            let looked_up = store.get_chunk(chunk_id);
-            if let Err(ref e) = looked_up {
-                note_store_failure(warnings, aliases, idx, "chunk lookup", e);
-            }
-            if let Ok(Some(chunk)) = looked_up {
-                return Some(crate::vectordb::SearchResult {
-                    id: chunk_id,
-                    content: chunk.content,
-                    path: chunk.path,
-                    start_line: chunk.start_line,
-                    end_line: chunk.end_line,
-                    kind: chunk.kind,
-                    signature: chunk.signature,
-                    docstring: chunk.docstring,
-                    context: chunk.context,
-                    hash: chunk.hash,
-                    distance: 0.0,
-                    score,
-                    context_prev: chunk.context_prev,
-                    context_next: chunk.context_next,
-                });
-            }
-        }
-        None
-    }
-
-    /// Resolve FTS results to SearchResult using multiple stores.
-    async fn resolve_fts_to_search_results_multi(
-        &self,
-        fts_results: &[crate::fts::FtsResult],
-        limit: usize,
-        stores: &[Arc<SharedStores>],
-        aliases: &[String],
-        warnings: &mut Vec<String>,
-    ) -> Vec<crate::vectordb::SearchResult> {
-        let mut results = Vec::new();
-        for fts in fts_results.iter().take(limit) {
-            for (idx, store_arc) in stores.iter().enumerate() {
-                let store = store_arc.vector_store.read().await;
-                let looked_up = store.get_chunk(fts.chunk_id);
-                if let Err(ref e) = looked_up {
-                    // `Ok(None)` means "this store does not hold that chunk" and
-                    // is normal during fan-out; `Err` means the store is broken.
-                    // Collapsing the two is how a dead vector store renders as
-                    // an empty literal search — the exact shape of the step-8
-                    // incident, which tantivy-side checks cannot detect.
-                    note_store_failure(warnings, aliases, idx, "chunk lookup", e);
-                }
-                if let Ok(Some(chunk)) = looked_up {
-                    results.push(crate::vectordb::SearchResult {
-                        id: fts.chunk_id,
-                        content: chunk.content,
-                        path: chunk.path,
-                        start_line: chunk.start_line,
-                        end_line: chunk.end_line,
-                        kind: chunk.kind,
-                        signature: chunk.signature,
-                        docstring: chunk.docstring,
-                        context: chunk.context,
-                        hash: chunk.hash,
-                        distance: 0.0,
-                        score: fts.score,
-                        context_prev: chunk.context_prev,
-                        context_next: chunk.context_next,
-                    });
-                    break; // Found in this store, skip remaining stores
-                }
-            }
-        }
-        results
-    }
-
-    /// Lexical-only search: FTS without embedding service.
-    #[allow(clippy::too_many_arguments)]
-    async fn semantic_search_lexical(
-        &self,
-        request: &SemanticSearchRequest,
-        identifiers: &[String],
-        limit: usize,
-        compact: bool,
-        stores: Option<Arc<SharedStores>>,
-        project_alias: Option<&str>,
-        alias_roots: &std::collections::HashMap<String, String>,
-    ) -> Result<CallToolResult, McpError> {
-        let structural_intent = detect_structural_intent(&request.query);
-
-        // `project=`-scoped queries route here, not through the fan-out
-        // (`is_multi` requires >1 store), so this path needs the same failure
-        // reporting — it is at least as common as a group query.
-        let mut lexical_warnings: Vec<String> = Vec::new();
-
-        let mut fts_results = match self
-            .with_fts_store_read_for(
-                |fts_store| fts_store.search(&request.query, limit * 5, structural_intent),
-                stores.clone(),
-            )
-            .await
-        {
-            Ok(r) => r,
-            Err(e) => {
-                let msg = format!("literal search failed: {e:#}");
-                tracing::error!("MCP: {}", msg);
-                lexical_warnings.push(msg);
-                Vec::new()
-            }
-        };
-
-        // Also do exact search if identifiers detected
-        for ident in identifiers {
-            let exact = match self
-                .with_fts_store_read_for(
-                    |fts_store| fts_store.search_exact(ident, limit * 3, structural_intent),
-                    stores.clone(),
-                )
-                .await
-            {
-                Ok(r) => r,
-                Err(e) => {
-                    let msg = format!("exact-identifier search for '{ident}' failed: {e:#}");
-                    tracing::error!("MCP: {}", msg);
-                    lexical_warnings.push(msg);
-                    continue;
-                }
-            };
-            merge_exact_into_fts(&mut fts_results, exact);
-        }
-
-        fts_results.sort_by(|a, b| {
-            b.score
-                .partial_cmp(&a.score)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        // Resolve FTS results to chunk metadata
-        let mut results = self
-            .resolve_fts_to_search_results(&fts_results, limit, stores, &mut lexical_warnings)
-            .await;
-
-        // Apply kind boost
-        if let Some(target_kind) = structural_intent {
-            boost_kind(&mut results, target_kind);
-        }
-
-        self.build_semantic_response(
-            results,
-            request,
-            compact,
-            !identifiers.is_empty(),
-            project_alias,
-            alias_roots,
-            &lexical_warnings,
-        )
-    }
-
-    /// Build the final SemanticSearchResponse with low-confidence signaling.
-    // Eight parameters, one over clippy's threshold. Bundling them into a
-    // `ResponseContext` struct is the right end state and is recorded as a
-    // follow-up; doing it in an incident fix would touch all seven call sites
-    // for no behavioural gain. The alternative — dropping `warnings` — is not
-    // acceptable: without it a failed repo is silently reported as "no match".
-    #[allow(clippy::too_many_arguments)]
-    fn build_semantic_response(
-        &self,
-        results: Vec<crate::vectordb::SearchResult>,
-        request: &SemanticSearchRequest,
-        compact: bool,
-        has_identifiers: bool,
-        project_alias: Option<&str>,
-        alias_roots: &std::collections::HashMap<String, String>,
-        // Repos that failed during a fan-out. MUST reach the caller: the
-        // consumer of this tool is a remote agent that never sees the server
-        // log, so a silently omitted repo reads as "no match there" — a false
-        // negative. The federated path already does this (`warnings` on the
-        // remote-project fan-out); the local path never could.
-        warnings: &[String],
-    ) -> Result<CallToolResult, McpError> {
-        let warnings = if warnings.is_empty() {
-            None
-        } else {
-            Some(warnings.to_vec())
-        };
-        if results.is_empty() {
-            let response = SemanticSearchResponse {
-                results: vec![],
-                low_confidence: Some(true),
-                suggested_tool: retry_hint(Some("literal_search".to_string()), &warnings),
-                warnings,
-            };
-            let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-            return Ok(CallToolResult::success(vec![Content::text(json)]));
-        }
-
-        // Pre-compute normalized project root for stripping absolute paths
-        let project_root_normalized = {
-            let root = crate::cache::normalize_path_str(self.project_path.to_str().unwrap_or(""));
-            root.trim_end_matches('/').to_string()
-        };
-
-        let mut items: Vec<SearchResultItem> = results
-            .into_iter()
-            .filter(|r| {
-                if let Some(ref fp) = request.filter_path {
-                    let normalized_filter = crate::cache::normalize_filter_path(fp);
-                    if normalized_filter.is_empty() {
-                        return true;
-                    }
-                    // Relativise against the ROUTED project's root, not the
-                    // service's own project_path — otherwise a serve-routed
-                    // absolute path never strips and every hit is dropped.
-                    let filter_root = pick_filter_root(
-                        &r.path,
-                        project_alias,
-                        alias_roots,
-                        &project_root_normalized,
-                    );
-                    crate::cache::path_matches_filter(&r.path, &normalized_filter, &filter_root)
-                } else {
-                    true
-                }
-            })
-            .map(|r| SearchResultItem {
-                chunk_id: Some(r.id),
-                path: r.path,
-                start_line: r.start_line,
-                end_line: r.end_line,
-                kind: r.kind,
-                score: r.score,
-                signature: r.signature,
-                content: if compact { None } else { Some(r.content) },
-                context_prev: if compact { None } else { r.context_prev },
-                context_next: if compact { None } else { r.context_next },
-                source: None,
-                chunk_ref: None,
-            })
-            .collect();
-
-        // Prefix paths with alias for multi-repo / single-project identification
-        for item in &mut items {
-            if let Some(alias) = project_alias {
-                if let Some(root) = alias_roots.get(alias) {
-                    item.path = prefix_path_with_alias(&item.path, Some(alias), root);
-                } else {
-                    item.path = crate::cache::normalize_path_str(&item.path);
-                }
-            } else if !alias_roots.is_empty() {
-                item.path = prefix_path_multi(&item.path, &[], alias_roots);
-            }
-        }
-
-        // Check low-confidence: top result's RRF score below threshold
-        let top_score = items.first().map(|r| r.score);
-        let (low_confidence, suggested_tool) = compute_low_confidence(top_score, has_identifiers);
-        let suggested_tool = retry_hint(suggested_tool, &warnings);
-
-        let response = SemanticSearchResponse {
-            results: items,
-            low_confidence,
-            suggested_tool,
-            warnings,
-        };
-
-        let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-        Ok(CallToolResult::success(vec![Content::text(json)]))
-    }
-
-    /// Resolve FTS results to SearchResult by looking up chunk metadata.
-    async fn resolve_fts_to_search_results(
-        &self,
-        fts_results: &[crate::fts::FtsResult],
-        limit: usize,
-        stores: Option<Arc<SharedStores>>,
-        warnings: &mut Vec<String>,
-    ) -> Vec<crate::vectordb::SearchResult> {
-        let outcome = self
-            .with_vector_store_read_for(
-                |store| {
-                    let mut results = Vec::new();
-                    for fts in fts_results.iter().take(limit) {
-                        // A failed lookup is not an absent chunk. Propagating the
-                        // error keeps a broken vector store from rendering as an
-                        // ordinary empty literal search.
-                        let chunk = store
-                            .get_chunk(fts.chunk_id)
-                            .context("Error resolving FTS hit to chunk metadata")?;
-                        if let Some(chunk) = chunk {
-                            results.push(crate::vectordb::SearchResult {
-                                id: fts.chunk_id,
-                                content: chunk.content,
-                                path: chunk.path,
-                                start_line: chunk.start_line,
-                                end_line: chunk.end_line,
-                                kind: chunk.kind,
-                                signature: chunk.signature,
-                                docstring: chunk.docstring,
-                                context: chunk.context,
-                                hash: chunk.hash,
-                                distance: 0.0,
-                                score: fts.score,
-                                context_prev: chunk.context_prev,
-                                context_next: chunk.context_next,
-                            });
-                        }
-                    }
-                    Ok(results)
-                },
-                stores,
-            )
-            .await;
-        match outcome {
-            Ok(results) => results,
-            Err(e) => {
-                let msg = format!("literal search could not read the index: {e:#}");
-                tracing::error!("MCP: {}", msg);
-                if !warnings.contains(&msg) {
-                    warnings.push(msg);
-                }
-                Vec::new()
-            }
-        }
-    }
-
-    // === find_definition internal ===
-
-    /// Internal: find symbol definitions, used by `find(kind="definition")`.
-    async fn find_definition(
-        &self,
-        Parameters(request): Parameters<FindDefinitionRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        let limit = request.limit.unwrap_or(20);
-
-        tracing::debug!(
-            "MCP find_definition: symbol='{}', kind={:?}, limit={}",
-            request.symbol,
-            request.kind,
-            limit
-        );
-
-        // Resolve project/group routing
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "find")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        // Stores that failed during this lookup. Without this, "the symbol may
-        // not be indexed" below is emitted as a confident diagnosis even when
-        // no store ever answered.
-        let mut find_warnings: Vec<String> = Vec::new();
-
-        // FTS search — multi-store or single
-        let fts_results = if let Some(ref sv) = ctx.stores_vec {
-            let sa = ctx.store_aliases.as_ref().unwrap();
-            self.with_fts_store_read_multi(
-                |fts_store| fts_store.search(&request.symbol, limit * 3, None),
-                sv.clone(),
-                sa,
-            )
-            .await
-            .unwrap_or_default()
-            .into_results(&mut find_warnings, "definition search")
-        } else {
-            match self
-                .with_fts_store_read_for(
-                    |fts_store| fts_store.search(&request.symbol, limit * 3, None),
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(r) => r,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error searching: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        if fts_results.is_empty() {
-            return Ok(CallToolResult::success(vec![Content::text(
-                qualify_empty_result(
-                    format!(
-                        "No definition found for '{}'. The symbol may not be indexed.",
-                        request.symbol
-                    ),
-                    &find_warnings,
-                ),
-            )]));
-        }
-
-        // Resolve chunk metadata and filter by definition kinds
-        let requested_kind = request.kind.clone();
-        let mut items: Vec<ReferenceItem> = if let Some(ref sv) = ctx.stores_vec {
-            let aliases = ctx.aliases();
-            let mut items: Vec<ReferenceItem> = Vec::new();
-            'outer: for fts_result in &fts_results {
-                for (store_idx, store_arc) in sv.iter().enumerate() {
-                    let store = store_arc.vector_store.read().await;
-                    let looked_up = store.get_chunk(fts_result.chunk_id);
-                    if let Err(ref e) = looked_up {
-                        // `Ok(None)` = chunk not in this store (normal during
-                        // fan-out); `Err` = broken store. Skipping the `Err`
-                        // silently made a dead store look like "symbol not
-                        // found" — carry it in the warnings channel instead.
-                        note_store_failure(
-                            &mut find_warnings,
-                            aliases,
-                            store_idx,
-                            "chunk lookup",
-                            e,
-                        );
-                    }
-                    if let Ok(Some(chunk)) = looked_up {
-                        // Skip non-definition kinds — try next FTS result, not next store
-                        if !DEFINITION_KINDS.contains(&chunk.kind.as_str()) {
-                            continue 'outer;
-                        }
-                        if let Some(ref rk) = requested_kind {
-                            if chunk.kind != *rk {
-                                continue 'outer;
-                            }
-                        }
-                        items.push(ReferenceItem {
-                            chunk_id: fts_result.chunk_id,
-                            path: chunk.path,
-                            line: chunk.start_line,
-                            kind: chunk.kind,
-                            signature: chunk.signature,
-                            score: fts_result.score,
-                        });
-                        if items.len() >= limit {
-                            break 'outer;
-                        }
-                        break; // Found in this store — move to next FTS result
-                    }
-                }
-                // If we get here, the chunk was Ok(None) in every store (not
-                // held anywhere — skip it) or its lookups failed (noted in
-                // find_warnings above).
-            }
-            items
-        } else {
-            match self
-                .with_vector_store_read_for(
-                    |store| {
-                        // Resolve chunk metadata first: a store `Err` must
-                        // reach the error arm below ("Error opening database")
-                        // instead of masquerading as a non-definition or
-                        // missing chunk — `Ok(None)` alone is a true miss.
-                        let resolved: anyhow::Result<Vec<_>> = fts_results
-                            .iter()
-                            .map(|fts_result| {
-                                let chunk = store.get_chunk(fts_result.chunk_id)?;
-                                Ok((chunk, fts_result.chunk_id, fts_result.score))
-                            })
-                            .collect();
-                        let items = resolved?
-                            .into_iter()
-                            .filter_map(|(looked_up, chunk_id, score)| {
-                                let chunk = looked_up?;
-                                if !DEFINITION_KINDS.contains(&chunk.kind.as_str()) {
-                                    return None;
-                                }
-                                if let Some(ref requested_kind) = requested_kind {
-                                    if chunk.kind != *requested_kind {
-                                        return None;
-                                    }
-                                }
-                                Some(ReferenceItem {
-                                    chunk_id,
-                                    path: chunk.path,
-                                    line: chunk.start_line,
-                                    kind: chunk.kind,
-                                    signature: chunk.signature,
-                                    score,
-                                })
-                            })
-                            .take(limit)
-                            .collect();
-                        Ok(items)
-                    },
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(items) => items,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error opening database: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        // Prefix paths with alias for multi-repo identification
-        for item in &mut items {
-            item.path = ctx.prefix_result_path(&item.path);
-        }
-
-        respond_with_items(&items, &find_warnings, || {
-            format!(
-                "No definition found for '{}'. Try find_usages() to find references, \
-                 or broaden your search.",
-                request.symbol
-            )
-        })
-    }
-
-    // === find_usages tool ===
-
-    async fn find_usages(
-        &self,
-        Parameters(request): Parameters<FindUsagesRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        self.find_usages_impl(
-            request.symbol.clone(),
-            request.limit.unwrap_or(20),
-            request.project,
-            request.group,
-        )
-        .await
-    }
-
-    /// Shared implementation for find_usages (used by `find(kind="usages")`).
-    async fn find_usages_impl(
-        &self,
-        symbol: String,
-        limit: usize,
-        project: Option<String>,
-        group: Option<String>,
-    ) -> Result<CallToolResult, McpError> {
-        tracing::debug!("MCP find_usages: symbol='{}', limit={}", symbol, limit);
-
-        // Resolve project/group routing
-        let ctx = match self.resolve_routing(&project, &group, false, "find").await {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        // See `find_definition`: an empty result and a dead store must not
-        // produce the same sentence.
-        let mut find_warnings: Vec<String> = Vec::new();
-
-        // FTS search — multi-store or single
-        let fts_results = if let Some(ref sv) = ctx.stores_vec {
-            let sa = ctx.store_aliases.as_ref().unwrap();
-            self.with_fts_store_read_multi(
-                |fts_store| fts_store.search(&symbol, limit * 2, None),
-                sv.clone(),
-                sa,
-            )
-            .await
-            .unwrap_or_default()
-            .into_results(&mut find_warnings, "usage search")
-        } else {
-            match self
-                .with_fts_store_read_for(
-                    |fts_store| fts_store.search(&symbol, limit * 2, None),
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(r) => r,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error searching: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        if fts_results.is_empty() {
-            return Ok(CallToolResult::success(vec![Content::text(
-                qualify_empty_result(
-                    format!("No usages found for '{symbol}'. The symbol may not be indexed."),
-                    &find_warnings,
-                ),
-            )]));
-        }
-
-        // Resolve chunks and exclude definition chunks
-        let mut items: Vec<ReferenceItem> = if let Some(ref sv) = ctx.stores_vec {
-            let aliases = ctx.aliases();
-            let mut items: Vec<ReferenceItem> = Vec::new();
-            for fts_result in &fts_results {
-                for (store_idx, store_arc) in sv.iter().enumerate() {
-                    let store = store_arc.vector_store.read().await;
-                    let looked_up = store.get_chunk(fts_result.chunk_id);
-                    if let Err(ref e) = looked_up {
-                        // Same rule as find_definition: `Err` is a broken
-                        // store, not "no usages" — carry it in the channel.
-                        note_store_failure(
-                            &mut find_warnings,
-                            aliases,
-                            store_idx,
-                            "chunk lookup",
-                            e,
-                        );
-                    }
-                    if let Ok(Some(chunk)) = looked_up {
-                        if !is_definition_chunk(&chunk.kind, &chunk.signature, &symbol) {
-                            items.push(ReferenceItem {
-                                chunk_id: fts_result.chunk_id,
-                                path: chunk.path,
-                                line: chunk.start_line,
-                                kind: chunk.kind,
-                                signature: chunk.signature,
-                                score: fts_result.score,
-                            });
-                        }
-                        break;
-                    }
-                }
-            }
-            items
-        } else {
-            match self
-                .with_vector_store_read_for(
-                    |store| {
-                        // Resolve first: a store `Err` must reach the error
-                        // arm below instead of masquerading as a definition
-                        // chunk or a miss — `Ok(None)` alone is a true miss.
-                        let resolved: anyhow::Result<Vec<_>> = fts_results
-                            .iter()
-                            .map(|fts_result| {
-                                let chunk = store.get_chunk(fts_result.chunk_id)?;
-                                Ok((chunk, fts_result.chunk_id, fts_result.score))
-                            })
-                            .collect();
-                        let items = resolved?
-                            .into_iter()
-                            .filter_map(|(looked_up, chunk_id, score)| {
-                                let chunk = looked_up?;
-                                if is_definition_chunk(&chunk.kind, &chunk.signature, &symbol) {
-                                    return None;
-                                }
-                                Some(ReferenceItem {
-                                    chunk_id,
-                                    path: chunk.path,
-                                    line: chunk.start_line,
-                                    kind: chunk.kind,
-                                    signature: chunk.signature,
-                                    score,
-                                })
-                            })
-                            .collect();
-                        Ok(items)
-                    },
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(items) => items,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error opening database: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        // Prefix paths with alias for multi-repo identification
-        for item in &mut items {
-            item.path = ctx.prefix_result_path(&item.path);
-        }
-
-        // Lexical FTS ranks docs, comments and code by the same text score,
-        // so a usages query can bury the real call-sites under markdown. The
-        // cut to `limit` MUST happen after the re-order: BM25 systematically
-        // scores short markdown blocks above long source files (document
-        // length normalisation), so ranking after a `take(limit)` would have
-        // nothing left to reorder exactly when it matters most. Stable sort
-        // preserves score order within each group; nothing is filtered.
-        rank_code_first(&mut items);
-        items.truncate(limit);
-
-        // When the hits include SCIP-backed source files and a precise
-        // backend is installed, tell the agent the exact upgrade path —
-        // otherwise the lexical list silently stands in for real references.
-        let note = scip_usages_note(&self.symbol_registry, &items, &symbol);
-
-        respond_with_items_noted(&items, &find_warnings, note.as_deref(), || {
-            format!(
-                "No usages found for '{symbol}' (only definitions were found). Try \
-                 find_definition() to locate the declaration."
-            )
-        })
-    }
-
-    /// Fetch outline items for an already-normalised absolute path.
-    ///
-    /// Returns `Ok(vec![])` when no chunks match.
-    /// In multi-store mode, per-store I/O failures are recorded in `warnings` and
-    /// skipped (never `Err`) so one broken repo cannot blank the whole outline.
-    /// In single-store mode, I/O failures are returned as `Err`.
-    ///
-    /// `warnings` is not optional: without it a failed store is indistinguishable
-    /// from a file with no indexed chunks, and the caller is told the file is not
-    /// indexed — a diagnosis, and a wrong one.
-    async fn outline_items_for_normalized(
-        &self,
-        normalized: &str,
-        ctx: &MultiStoreContext,
-        warnings: &mut Vec<String>,
-    ) -> anyhow::Result<Vec<FileOutlineItem>> {
-        if let Some(ref sv) = ctx.stores_vec {
-            let aliases = ctx.aliases();
-            let mut all_items: Vec<FileOutlineItem> = Vec::new();
-            let mut seen_ids: std::collections::HashSet<u32> = std::collections::HashSet::new();
-            for (store_idx, store_arc) in sv.iter().enumerate() {
-                let store = store_arc.vector_store.read().await;
-                match store.chunks_for_file(normalized) {
-                    Ok(metas) => {
-                        for c in metas {
-                            if seen_ids.insert(c.id) {
-                                all_items.push(FileOutlineItem {
-                                    chunk_id: c.id,
-                                    kind: c.kind,
-                                    signature: c.signature,
-                                    start_line: c.start_line,
-                                    end_line: c.end_line,
-                                });
-                            }
-                        }
-                    }
-                    Err(ref e) => {
-                        note_store_failure(warnings, aliases, store_idx, "outline scan", e);
-                    }
-                }
-            }
-            all_items.sort_by_key(|i| i.start_line);
-            Ok(all_items)
-        } else {
-            let normalized_owned = normalized.to_string();
-            self.with_vector_store_read_for(
-                move |store| {
-                    let mut out: Vec<FileOutlineItem> = store
-                        .chunks_for_file(&normalized_owned)?
-                        .into_iter()
-                        .map(|c| FileOutlineItem {
-                            chunk_id: c.id,
-                            kind: c.kind,
-                            signature: c.signature,
-                            start_line: c.start_line,
-                            end_line: c.end_line,
-                        })
-                        .collect();
-                    out.sort_by_key(|i| i.start_line);
-                    Ok(out)
-                },
-                ctx.stores.clone(),
-            )
-            .await
-        }
-    }
-
-    async fn file_outline(
-        &self,
-        Parameters(request): Parameters<FileOutlineRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        // Resolve project/group routing
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "explore")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        // Outline operates on a single repo — reject group fan-out
-        if ctx.is_multi {
-            return Ok(CallToolResult::success(vec![Content::text(
-                "Tool 'explore' operates on a single repo. Use 'project' instead of 'group'."
-                    .to_string(),
-            )]));
-        }
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        // In serve mode, use the resolved project root from alias_roots;
-        // self.project_path is "serve://multi-repo" which doesn't resolve.
-        let project_root = if let Some(ref alias) = ctx.project_alias {
-            ctx.alias_roots
-                .get(alias)
-                .map(PathBuf::from)
-                .unwrap_or_else(|| self.project_path.clone())
-        } else {
-            self.project_path.clone()
-        };
-        // Strip project-alias prefix from target path if present.
-        // E.g. "ExampleRepo/src/foo.cs" with project="ExampleRepo" → "src/foo.cs"
-        let stripped_path = strip_alias_prefix(&request.path, ctx.project_alias.as_ref());
-        let normalized = normalize_tool_path(&stripped_path, &project_root);
-
-        let mut outline_warnings: Vec<String> = Vec::new();
-        let mut items = match self
-            .outline_items_for_normalized(&normalized, &ctx, &mut outline_warnings)
-            .await
-        {
-            Ok(v) => v,
-            Err(e) => {
-                return Ok(CallToolResult::success(vec![Content::text(format!(
-                    "Error reading outline: {e:#}"
-                ))]));
-            }
-        };
-
-        // Two-pass fallback: if alias-stripping changed the path and yielded no results,
-        // try the original un-stripped path. Handles the case where the project alias
-        // matches a package subdirectory name (e.g. project "my_pkg" with target
-        // "my_pkg/config.py" → after strip becomes "config.py" which is wrong;
-        // the correct relative path is "my_pkg/config.py").
-        if items.is_empty() && stripped_path != request.path {
-            let normalized_orig = normalize_tool_path(&request.path, &project_root);
-            if normalized_orig != normalized {
-                tracing::debug!(
-                    "file_outline: primary '{}' empty, trying fallback '{}'",
-                    normalized,
-                    normalized_orig
-                );
-                items = match self
-                    .outline_items_for_normalized(&normalized_orig, &ctx, &mut outline_warnings)
-                    .await
-                {
-                    Ok(v) => v,
-                    Err(e) => {
-                        tracing::warn!(
-                            "file_outline: fallback '{}' also failed: {:?}",
-                            normalized_orig,
-                            e
-                        );
-                        push_store_warning(
-                            &mut outline_warnings,
-                            &store_warning(
-                                ctx.project_alias.as_deref().unwrap_or("unknown"),
-                                "outline scan",
-                                &format!("{e:#}"),
-                            ),
-                        );
-                        Vec::new()
-                    }
-                };
-            }
-        }
-
-        respond_with_items(&items, &outline_warnings, || {
-            "No indexed chunks found for path. Verify the file is within the \
-             project root and the index is up to date."
-                .to_string()
-        })
-    }
-
-    #[tool(
-        description = "Retrieve the full content of a specific chunk by its ID, plus optional surrounding lines for context.\nUse this after search or explore to read the actual code without loading the whole file.\n\nUSE FOR: reading a specific function/class body after finding it via search.\nSet context_lines (default 0, max 20) to include lines before and after the chunk.\n\nIMPORTANT (multi-repo): chunk_ids are local to each repository and are NOT globally unique.\nWhen `project` is omitted in multi-repo mode, the tool scans all repositories for the chunk_id.\nIf found in exactly one repo, it is returned automatically. If found in multiple repos, an `ambiguous_chunk_id` error lists the candidates so you can retry with `project`."
-    )]
-    async fn get_chunk(
-        &self,
-        Parameters(request): Parameters<GetChunkRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        tracing::info!(
-            "📥 get_chunk(chunk_id={}, project={:?})",
-            request.chunk_id,
-            request.project,
-        );
-
-        // Federation: a `chunk_ref` of the form "<peer>/<remote_alias>:<chunk_id>"
-        // (returned by a federated search result) fetches the chunk from a remote
-        // peer rather than the local index. The alias scopes the fetch to a single
-        // remote project so the multi-repo peer can disambiguate the chunk_id.
-        if let Some(chunk_ref) = request.chunk_ref.as_deref() {
-            return self
-                .federated_get_chunk(chunk_ref, request.context_lines)
-                .await;
-        }
-
-        // In multi-repo serve mode, require explicit project or group scope.
-        // Unscoped get_chunk would fan-out over all repos, opening all DBs unnecessarily.
-        // Consistent with search/find/explore which also require scope.
-        if request.project.is_none() && request.group.is_none() {
-            if let Some(ref serve_state) = self.serve_state {
-                let config = serve_state.config_snapshot();
-                if config.repos.len() > 1 {
-                    return Ok(CallToolResult::success(vec![Content::text(
-                        self.format_scope_error(),
-                    )]));
-                }
-            }
-        }
-
-        // Resolve project/group routing — allow unscoped only for single-repo mode
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, true, "get_chunk")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        let mut clamped = false;
-        let mut context_lines = request.context_lines.unwrap_or(0);
-        if context_lines > 20 {
-            context_lines = 20;
-            clamped = true;
-        }
-
-        // Stores that failed while looking up this chunk. get_chunk previously
-        // collapsed every `Err` into "not found", so during the read-only
-        // incident it would have reported every chunk in every vendor repo as
-        // missing — a confident, wrong answer.
-        let mut chunk_warnings: Vec<String> = Vec::new();
-
-        // Look up chunk — multi-store: smart candidate detection for chunk_id collision.
-        // chunk_ids are local per database, not globally unique. When no project is specified
-        // and multiple stores are active, scan all stores to find which ones have this chunk_id.
-        let chunk = if let Some(ref sv) = ctx.stores_vec {
-            if sv.len() > 1 && request.project.is_none() {
-                // Smart candidate detection: find which stores actually contain this chunk_id
-                let mut candidates: Vec<(&Arc<SharedStores>, String)> = Vec::new();
-                let aliases = ctx.aliases();
-                for (i, store_arc) in sv.iter().enumerate() {
-                    let store = store_arc.vector_store.read().await;
-                    match store.get_chunk(request.chunk_id) {
-                        Ok(Some(_)) => {
-                            // A store that HAS the chunk stays a candidate even if
-                            // its alias is missing. `resolve_repo_stores_multi`
-                            // keeps stores and aliases the same length, so this is
-                            // unreachable today — but gating the push on
-                            // `aliases.get(i)` meant a future break of that
-                            // invariant would degrade to a silent auto-route rather
-                            // than a loud one. The placeholder is per-index so two
-                            // aliasless candidates stay distinguishable in
-                            // `candidate_projects`.
-                            let alias = aliases
-                                .get(i)
-                                .cloned()
-                                .unwrap_or_else(|| format!("<unnamed store #{i}>"));
-                            candidates.push((store_arc, alias));
-                        }
-                        Ok(None) => continue,
-                        Err(ref e) => {
-                            note_store_failure(&mut chunk_warnings, aliases, i, "chunk lookup", e);
-                            continue;
-                        }
-                    }
-                }
-                match candidates.len() {
-                    0 => {
-                        return Ok(CallToolResult::success(vec![Content::text(
-                            qualify_empty_result(
-                                format!(
-                                    "Chunk {} not found in any repository. Verify the \
-                                     chunk_id and index state.",
-                                    request.chunk_id
-                                ),
-                                &chunk_warnings,
-                            ),
-                        )]));
-                    }
-                    1 => {
-                        // Exactly one store has this chunk_id — auto-route
-                        let (store_arc, ref alias) = candidates[0];
-                        // Record tool call for the specific repo that served this chunk
-                        if let Some(ref serve_state) = self.serve_state {
-                            serve_state.record_tool_call(alias, "get_chunk");
-                            serve_state.touch_access(alias);
-                        }
-                        let store = store_arc.vector_store.read().await;
-                        match store.get_chunk(request.chunk_id) {
-                            Ok(c) => c,
-                            Err(ref e) => {
-                                push_store_warning(
-                                    &mut chunk_warnings,
-                                    &store_warning(alias, "chunk lookup", &format!("{e:#}")),
-                                );
-                                None
-                            }
-                        }
-                    }
-                    _ => {
-                        // Multiple stores have this chunk_id — ambiguous.
-                        //
-                        // `candidate_projects` reads as the complete list, so a
-                        // store that failed to answer has to be declared: the
-                        // right repo may be the one missing from it.
-                        let candidate_names: Vec<&str> =
-                            candidates.iter().map(|(_, a)| a.as_str()).collect();
-                        let payload = ambiguous_chunk_payload(
-                            request.chunk_id,
-                            &candidate_names,
-                            &chunk_warnings,
-                        );
-                        return Ok(CallToolResult::success(vec![Content::text(
-                            payload.to_string(),
-                        )]));
-                    }
-                }
-            } else {
-                // Single store or project specified — direct lookup
-                let aliases = ctx.aliases();
-                let mut found = None;
-                for (i, store_arc) in sv.iter().enumerate() {
-                    let store = store_arc.vector_store.read().await;
-                    match store.get_chunk(request.chunk_id) {
-                        Ok(Some(c)) => {
-                            found = Some(c);
-                            break;
-                        }
-                        Ok(None) => continue,
-                        // Do NOT abandon the remaining stores: one broken store
-                        // says nothing about the others, and the chunk may well
-                        // live in a healthy one.
-                        Err(ref e) => {
-                            note_store_failure(&mut chunk_warnings, aliases, i, "chunk lookup", e);
-                            continue;
-                        }
-                    }
-                }
-                found
-            }
-        } else {
-            match self
-                .with_vector_store_read_for(
-                    |store| store.get_chunk(request.chunk_id),
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(c) => c,
-                Err(e) => {
-                    push_store_warning(
-                        &mut chunk_warnings,
-                        &store_warning(
-                            ctx.project_alias.as_deref().unwrap_or("unknown"),
-                            "chunk lookup",
-                            &format!("{e:#}"),
-                        ),
-                    );
-                    None
-                }
-            }
-        };
-
-        let mut chunk = match chunk {
-            Some(c) => c,
-            None => {
-                return Ok(CallToolResult::success(vec![Content::text(
-                    qualify_empty_result(
-                        format!(
-                            "Chunk {} not found. Verify the chunk_id and index state.",
-                            request.chunk_id
-                        ),
-                        &chunk_warnings,
-                    ),
-                )]));
-            }
-        };
-
-        // Prefix path with alias for multi-repo identification
-        chunk.path = ctx.prefix_result_path(&chunk.path);
-
-        let mut context_before = None;
-        let mut context_after = None;
-        let mut note = None;
-
-        if context_lines > 0 {
-            // Resolve relative chunk paths against project root (not process CWD).
-            let source_path = if Path::new(&chunk.path).is_absolute() {
-                PathBuf::from(&chunk.path)
-            } else {
-                self.project_path.join(&chunk.path)
-            };
-            match tokio::fs::read_to_string(&source_path).await {
-                Ok(src) => {
-                    let lines: Vec<&str> = src.lines().collect();
-                    if !lines.is_empty() {
-                        let before_start = chunk.start_line.saturating_sub(context_lines);
-                        let before_end = chunk.start_line.min(lines.len());
-                        if before_start < before_end {
-                            context_before = Some(lines[before_start..before_end].join("\n"));
-                        }
-
-                        let after_start = chunk.end_line.min(lines.len());
-                        let after_end = (chunk.end_line + context_lines).min(lines.len());
-                        if after_start < after_end {
-                            context_after = Some(lines[after_start..after_end].join("\n"));
-                        }
-                    }
-                }
-                Err(_) => {
-                    note = Some(
-                        "source file not readable, returning indexed content only".to_string(),
-                    );
-                }
-            }
-        }
-
-        let response = GetChunkResponse {
-            chunk_id: request.chunk_id,
-            path: chunk.path,
-            start_line: chunk.start_line,
-            end_line: chunk.end_line,
-            kind: chunk.kind,
-            signature: chunk.signature,
-            content: chunk.content,
-            context_before,
-            context_after,
-            context_lines_clamped: if clamped { Some(true) } else { None },
-            note,
-        };
-
-        // The success path is the one that used to drop this, and it is the
-        // dangerous one: a confidently-returned chunk from a group where a store
-        // failed to answer looks exactly like a chunk from a healthy group. Same
-        // false negative as an empty result, harder to notice.
-        respond_with_object(&response, &chunk_warnings)
-    }
-
-    fn normalize_symbol_query_path(&self, project_root: &Path, file: &Path) -> PathBuf {
-        if file.is_absolute() {
-            if let Ok(relative) = file.strip_prefix(project_root) {
-                return PathBuf::from(relative.to_string_lossy().replace('\\', "/"));
-            }
-        }
-
-        PathBuf::from(file.to_string_lossy().replace('\\', "/"))
-    }
-
-    async fn find_imports(
-        &self,
-        Parameters(request): Parameters<FindImportsRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        // Resolve project/group routing
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "find")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        // In serve mode, use the resolved project root from alias_roots
-        let project_root = if let Some(ref alias) = ctx.project_alias {
-            ctx.alias_roots
-                .get(alias)
-                .map(PathBuf::from)
-                .unwrap_or_else(|| self.project_path.clone())
-        } else {
-            self.project_path.clone()
-        };
-        // Strip project-alias prefix from target path if present.
-        let stripped_path = strip_alias_prefix(&request.path, ctx.project_alias.as_ref());
-        let normalized = normalize_tool_path(&stripped_path, &project_root);
-
-        // Stores that failed during this lookup, so "no imports found" is never
-        // reported as fact when a store never answered.
-        let mut import_warnings: Vec<String> = Vec::new();
-
-        let mut items = if let Some(ref sv) = ctx.stores_vec {
-            // Multi-store group fan-out: collect import items from all stores
-            let import_aliases = ctx.aliases();
-            let mut all_items: Vec<ImportItem> = Vec::new();
-            let mut seen_ids: std::collections::HashSet<u32> = std::collections::HashSet::new();
-            for (store_idx, store_arc) in sv.iter().enumerate() {
-                let store = store_arc.vector_store.read().await;
-                match store.chunks_for_file(&normalized) {
-                    Ok(metas) => {
-                        for meta in metas {
-                            if !is_import_kind(&meta.kind) {
-                                continue;
-                            }
-                            if seen_ids.insert(meta.id) {
-                                match store.get_chunk(meta.id) {
-                                    Ok(Some(chunk)) => all_items.extend(parse_import_lines(
-                                        &chunk.content,
-                                        chunk.start_line,
-                                    )),
-                                    Ok(None) => {}
-                                    Err(ref e) => note_store_failure(
-                                        &mut import_warnings,
-                                        import_aliases,
-                                        store_idx,
-                                        "chunk lookup",
-                                        e,
-                                    ),
-                                }
-                            }
-                        }
-                    }
-                    Err(ref e) => {
-                        note_store_failure(
-                            &mut import_warnings,
-                            import_aliases,
-                            store_idx,
-                            "imports scan",
-                            e,
-                        );
-                    }
-                }
-            }
-            all_items
-        } else {
-            match self
-                .with_vector_store_read_for(
-                    |store| {
-                        let mut out = Vec::new();
-                        for meta in store.chunks_for_file(&normalized)? {
-                            if !is_import_kind(&meta.kind) {
-                                continue;
-                            }
-                            if let Some(chunk) = store.get_chunk(meta.id)? {
-                                out.extend(parse_import_lines(&chunk.content, chunk.start_line));
-                            }
-                        }
-                        Ok(out)
-                    },
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(items) => items,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error reading imports: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        if items.is_empty() {
-            // Fallback: no import-kind chunks found for this file. Broaden the
-            // search to common import keywords and filter to the target path.
-            // Limitation: this only finds chunks containing these literal words;
-            // language-specific import forms that lack these keywords will be missed.
-            let fallback_limit = 40usize;
-            let mut all_hits: Vec<(u32, f32)> = Vec::new();
-            let mut seen_fts_ids: HashSet<u32> = HashSet::new();
-
-            if let Some(ref sv) = ctx.stores_vec {
-                let import_aliases = ctx.aliases();
-                // Multi-store FTS fallback
-                for keyword in IMPORT_FTS_KEYWORDS {
-                    let hits = self
-                        .with_fts_store_read_multi(
-                            |fts_store| fts_store.search_exact(keyword, fallback_limit, None),
-                            sv.clone(),
-                            ctx.store_aliases.as_ref().unwrap(),
-                        )
-                        .await
-                        .unwrap_or_default()
-                        .into_results(&mut import_warnings, "imports search");
-                    for h in hits {
-                        if seen_fts_ids.insert(h.chunk_id) {
-                            all_hits.push((h.chunk_id, h.score));
-                        }
-                    }
-                }
-
-                // Resolve FTS hits via vector stores
-                let mut resolved: Vec<ImportItem> = Vec::new();
-                for (chunk_id, _) in &all_hits {
-                    for (store_idx, store_arc) in sv.iter().enumerate() {
-                        let store = store_arc.vector_store.read().await;
-                        match store.get_chunk(*chunk_id) {
-                            Ok(Some(chunk)) => {
-                                if crate::cache::normalize_path_str(&chunk.path) == normalized {
-                                    resolved.extend(parse_import_lines(
-                                        &chunk.content,
-                                        chunk.start_line,
-                                    ));
-                                }
-                                break;
-                            }
-                            Ok(None) => continue,
-                            Err(ref e) => {
-                                note_store_failure(
-                                    &mut import_warnings,
-                                    import_aliases,
-                                    store_idx,
-                                    "chunk lookup",
-                                    e,
-                                );
-                                continue;
-                            }
-                        }
-                    }
-                }
-                items = resolved;
-            } else {
-                // Single-store FTS fallback
-                for keyword in IMPORT_FTS_KEYWORDS {
-                    let hits = match self
-                        .with_fts_store_read_for(
-                            |fts_store| fts_store.search_exact(keyword, fallback_limit, None),
-                            ctx.stores.clone(),
-                        )
-                        .await
-                    {
-                        Ok(h) => h,
-                        Err(e) => {
-                            push_store_warning(
-                                &mut import_warnings,
-                                &store_warning(
-                                    ctx.project_alias.as_deref().unwrap_or("unknown"),
-                                    "imports search",
-                                    &format!("{e:#}"),
-                                ),
-                            );
-                            Vec::new()
-                        }
-                    };
-                    for h in hits {
-                        if seen_fts_ids.insert(h.chunk_id) {
-                            all_hits.push((h.chunk_id, h.score));
-                        }
-                    }
-                }
-
-                items = self
-                    .with_vector_store_read_for(
-                        |store| {
-                            let mut out = Vec::new();
-                            for (chunk_id, _) in &all_hits {
-                                if let Some(chunk) = store.get_chunk(*chunk_id)? {
-                                    if crate::cache::normalize_path_str(&chunk.path) == normalized {
-                                        out.extend(parse_import_lines(
-                                            &chunk.content,
-                                            chunk.start_line,
-                                        ));
-                                    }
-                                }
-                            }
-                            Ok(out)
-                        },
-                        ctx.stores.clone(),
-                    )
-                    .await
-                    .unwrap_or_else(|e| {
-                        push_store_warning(
-                            &mut import_warnings,
-                            &store_warning(
-                                ctx.project_alias.as_deref().unwrap_or("unknown"),
-                                "chunk lookup",
-                                &format!("{e:#}"),
-                            ),
-                        );
-                        Vec::new()
-                    });
-            }
-        }
-
-        items.sort_by_key(|i| i.line);
-        respond_with_items(&items, &import_warnings, || {
-            "No import chunks found. The index may not include import statements \
-             for this language, or the file has no imports."
-                .to_string()
-        })
-    }
-
-    async fn find_dependents(
-        &self,
-        Parameters(request): Parameters<FindDependentsRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        // Resolve project/group routing
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "find")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        let limit = request.limit.unwrap_or(20).min(200);
-        let high_limit = (limit * 10).max(200); // generous budget for filtering
-
-        // Stores that failed during this lookup, so "no dependents" is never
-        // reported as fact when a store never answered.
-        let mut dep_warnings: Vec<String> = Vec::new();
-
-        // Extract a meaningful search term from path-like inputs.
-        // Import chunks contain module references like `use crate::constants::X`
-        // but the tool receives file paths like `src/constants.rs`.
-        // We extract the file stem to match against module names in imports.
-        let search_term = if request.symbol_or_path.contains('/')
-            || request.symbol_or_path.contains('\\')
-            || request.symbol_or_path.contains('.')
-        {
-            std::path::Path::new(&request.symbol_or_path)
-                .file_stem()
-                .and_then(|s| s.to_str())
-                .filter(|s| !s.is_empty())
-                .unwrap_or(&request.symbol_or_path)
-                .to_string()
-        } else {
-            request.symbol_or_path.clone()
-        };
-
-        let import_kind = Some(crate::chunker::ChunkKind::Imports);
-
-        // Two-phase search strategy:
-        // 1. `search_exact` — precise term match on signature+content with
-        //    MUST filter for Import kind. Strictly limits results to import chunks.
-        // 2. If that yields no import-kind results, fall back to `search`
-        //    (QueryParser, broader tokenization) with kind boost for imports.
-        //
-        // Limitation: the chunker does not emit per-statement AST import chunks;
-        // imports are gap-classified as `Imports` kind. Chunks whose kind doesn't
-        // match `is_import_kind()` will be missed regardless of search method.
-        let fts_results = if let Some(ref sv) = ctx.stores_vec {
-            let sa = ctx.store_aliases.as_ref().unwrap();
-            // Multi-store FTS search
-            let exact_hits = self
-                .with_fts_store_read_multi(
-                    |fts_store| fts_store.search_exact(&search_term, high_limit, import_kind),
-                    sv.clone(),
-                    sa,
-                )
-                .await
-                .unwrap_or_default()
-                .into_results(&mut dep_warnings, "dependents search");
-
-            if exact_hits.is_empty() {
-                self.with_fts_store_read_multi(
-                    |fts_store| fts_store.search(&search_term, high_limit, import_kind),
-                    sv.clone(),
-                    sa,
-                )
-                .await
-                .unwrap_or_default()
-                .into_results(&mut dep_warnings, "dependents search")
-            } else {
-                exact_hits
-            }
-        } else {
-            // Single-store FTS search
-            let alias = ctx.project_alias.as_deref().unwrap_or("unknown");
-            let mut run = |r: anyhow::Result<Vec<crate::fts::FtsResult>>| match r {
-                Ok(hits) => hits,
-                Err(e) => {
-                    push_store_warning(
-                        &mut dep_warnings,
-                        &store_warning(alias, "dependents search", &format!("{e:#}")),
-                    );
-                    Vec::new()
-                }
-            };
-            let exact_hits = run(self
-                .with_fts_store_read_for(
-                    |fts_store| fts_store.search_exact(&search_term, high_limit, import_kind),
-                    ctx.stores.clone(),
-                )
-                .await);
-
-            if exact_hits.is_empty() {
-                run(self
-                    .with_fts_store_read_for(
-                        |fts_store| fts_store.search(&search_term, high_limit, import_kind),
-                        ctx.stores.clone(),
-                    )
-                    .await)
-            } else {
-                exact_hits
-            }
-        };
-
-        let mut items = if let Some(ref sv) = ctx.stores_vec {
-            // Multi-store: resolve chunks across all stores
-            let dep_aliases = ctx.aliases();
-            let mut seen_paths = HashSet::new();
-            let mut out = Vec::new();
-            for f in &fts_results {
-                for (store_idx, store_arc) in sv.iter().enumerate() {
-                    let store = store_arc.vector_store.read().await;
-                    match store.get_chunk(f.chunk_id) {
-                        Ok(Some(chunk)) => {
-                            if !is_import_kind(&chunk.kind) {
-                                break; // try next FTS result
-                            }
-
-                            let norm = crate::cache::normalize_path_str(&chunk.path);
-                            if !seen_paths.insert(norm) {
-                                break;
-                            }
-
-                            let term_lower = search_term.to_lowercase();
-                            let import_statement =
-                                if chunk.content.to_lowercase().contains(&term_lower) {
-                                    chunk
-                                        .content
-                                        .lines()
-                                        .find(|l| l.to_lowercase().contains(&term_lower))
-                                        .unwrap_or("")
-                                        .to_string()
-                                } else {
-                                    chunk.signature.filter(|s| !s.is_empty()).unwrap_or(
-                                        chunk.content.lines().next().unwrap_or("").to_string(),
-                                    )
-                                };
-
-                            out.push(DependentItem {
-                                path: chunk.path,
-                                line: chunk.start_line,
-                                import_statement,
-                            });
-
-                            break; // found in this store, move to next FTS result
-                        }
-                        Ok(None) => {} // try next store
-                        // One broken store says nothing about the others; a
-                        // `break` here silently drops a chunk that lives in a
-                        // healthy store later in the list.
-                        Err(ref e) => {
-                            note_store_failure(
-                                &mut dep_warnings,
-                                dep_aliases,
-                                store_idx,
-                                "chunk lookup",
-                                e,
-                            );
-                            continue;
-                        }
-                    }
-                }
-                if out.len() >= limit {
-                    break;
-                }
-            }
-            out
-        } else {
-            match self
-                .with_vector_store_read_for(
-                    |store| {
-                        let mut seen_paths = HashSet::new();
-                        let mut out = Vec::new();
-                        let term_lower = search_term.to_lowercase();
-                        for f in &fts_results {
-                            if let Some(chunk) = store.get_chunk(f.chunk_id)? {
-                                if !is_import_kind(&chunk.kind) {
-                                    continue;
-                                }
-
-                                let norm = crate::cache::normalize_path_str(&chunk.path);
-                                if !seen_paths.insert(norm) {
-                                    continue;
-                                }
-
-                                // Extract the specific import line(s) that mention the
-                                // module name, rather than returning the entire chunk content.
-                                let import_statement =
-                                    if chunk.content.to_lowercase().contains(&term_lower) {
-                                        chunk
-                                            .content
-                                            .lines()
-                                            .find(|l| l.to_lowercase().contains(&term_lower))
-                                            .unwrap_or("")
-                                            .to_string()
-                                    } else {
-                                        chunk.signature.filter(|s| !s.is_empty()).unwrap_or(
-                                            chunk.content.lines().next().unwrap_or("").to_string(),
-                                        )
-                                    };
-
-                                out.push(DependentItem {
-                                    path: chunk.path,
-                                    line: chunk.start_line,
-                                    import_statement,
-                                });
-
-                                if out.len() >= limit {
-                                    break;
-                                }
-                            }
-                        }
-                        Ok(out)
-                    },
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(items) => items,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error resolving dependents: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        // Prefix paths with alias for multi-repo identification
-        for item in &mut items {
-            item.path = ctx.prefix_result_path(&item.path);
-        }
-
-        items.sort_by(|a, b| a.path.cmp(&b.path));
-        respond_with_items(&items, &dep_warnings, || {
-            format!("No dependent files found for '{}'.", request.symbol_or_path)
-        })
-    }
-
-    /// Internal: find similar chunks, used by `explore(kind="similar")`.
-    async fn similar_chunks(
-        &self,
-        Parameters(request): Parameters<SimilarChunksRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        // Resolve project/group routing
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "explore")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        let limit = request.limit.unwrap_or(5).min(20);
-
-        // Stores that failed while resolving the source embedding. `if let
-        // Ok(Some(..))` used to discard the error, so a dead store produced
-        // "embedding not found" — a wrong diagnosis, not a missing chunk.
-        let mut similar_warnings: Vec<String> = Vec::new();
-
-        let mut results = if let Some(ref sv) = ctx.stores_vec {
-            // Multi-store: find the embedding in whichever store has it,
-            // then search across all stores for similar chunks.
-            let aliases = ctx.aliases();
-            let mut embedding: Option<Vec<f32>> = None;
-            for (i, store_arc) in sv.iter().enumerate() {
-                let store = store_arc.vector_store.read().await;
-                match store.get_embedding(request.chunk_id) {
-                    Ok(Some(emb)) => {
-                        embedding = Some(emb);
-                        break;
-                    }
-                    Ok(None) => continue,
-                    Err(ref e) => {
-                        note_store_failure(
-                            &mut similar_warnings,
-                            aliases,
-                            i,
-                            "embedding lookup",
-                            e,
-                        );
-                        continue;
-                    }
-                }
-            }
-
-            let embedding = match embedding {
-                Some(e) => e,
-                None => {
-                    return Ok(CallToolResult::success(vec![Content::text(
-                        qualify_empty_result(
-                            format!(
-                                "Embedding not found for chunk_id {} in any store.",
-                                request.chunk_id
-                            ),
-                            &similar_warnings,
-                        ),
-                    )]));
-                }
-            };
-
-            // Search across all stores with the found embedding
-            let mut all_results: Vec<SearchResultItem> = Vec::new();
-            let mut seen_ids: std::collections::HashSet<u32> = std::collections::HashSet::new();
-            for (store_idx, store_arc) in sv.iter().enumerate() {
-                let store = store_arc.vector_store.read().await;
-                match store.search(&embedding, limit + 1) {
-                    Ok(mut neighbors) => {
-                        neighbors.retain(|r| r.id != request.chunk_id);
-                        for r in neighbors {
-                            if seen_ids.insert(r.id) {
-                                all_results.push(SearchResultItem {
-                                    chunk_id: Some(r.id),
-                                    path: r.path,
-                                    start_line: r.start_line,
-                                    end_line: r.end_line,
-                                    kind: r.kind,
-                                    score: r.score,
-                                    signature: r.signature,
-                                    content: None,
-                                    context_prev: None,
-                                    context_next: None,
-                                    source: None,
-                                    chunk_ref: None,
-                                });
-                            }
-                        }
-                    }
-                    Err(ref e) => {
-                        // The embedding was found, so the handler returns results
-                        // either way; without this, a group query silently omits
-                        // every neighbour from the broken repo.
-                        note_store_failure(
-                            &mut similar_warnings,
-                            aliases,
-                            store_idx,
-                            "similarity search",
-                            e,
-                        );
-                    }
-                }
-            }
-
-            all_results.sort_by(|a, b| {
-                b.score
-                    .partial_cmp(&a.score)
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            });
-            all_results.truncate(limit);
-            all_results
-        } else {
-            match self
-                .with_vector_store_read_for(
-                    |store| {
-                        let embedding =
-                            store.get_embedding(request.chunk_id)?.ok_or_else(|| {
-                                anyhow::anyhow!(
-                                    "embedding not found for chunk_id {}",
-                                    request.chunk_id
-                                )
-                            })?;
-
-                        let mut neighbors = store.search(&embedding, limit + 1)?;
-                        neighbors.retain(|r| r.id != request.chunk_id);
-                        neighbors.truncate(limit);
-
-                        let items = neighbors
-                            .into_iter()
-                            .map(|r| SearchResultItem {
-                                chunk_id: Some(r.id),
-                                path: r.path,
-                                start_line: r.start_line,
-                                end_line: r.end_line,
-                                kind: r.kind,
-                                score: r.score,
-                                signature: r.signature,
-                                content: None,
-                                context_prev: None,
-                                context_next: None,
-                                source: None,
-                                chunk_ref: None,
-                            })
-                            .collect::<Vec<_>>();
-                        Ok(items)
-                    },
-                    ctx.stores.clone(),
-                )
-                .await
-            {
-                Ok(items) => items,
-                Err(e) => {
-                    return Ok(CallToolResult::success(vec![Content::text(format!(
-                        "Error finding similar chunks: {e:#}"
-                    ))]));
-                }
-            }
-        };
-
-        // Prefix paths with alias for multi-repo identification
-        for item in &mut results {
-            item.path = ctx.prefix_result_path(&item.path);
-        }
-
-        // Every exit carries the channel: the earlier read sat in an
-        // early-return arm, so once an embedding was found, every failure
-        // recorded afterwards (the whole neighbour fan-out) was discarded.
-        respond_with_items(&results, &similar_warnings, || {
-            format!("No similar chunks found for chunk_id {}.", request.chunk_id)
-        })
-    }
-
-    async fn literal_search(
-        &self,
-        Parameters(request): Parameters<LiteralSearchRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        // Resolve project/group routing
-        let ctx = match self
-            .resolve_routing(&request.project, &request.group, false, "search")
-            .await
-        {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        let limit = request.limit.unwrap_or(20);
-        let output_format = request.format.as_deref().unwrap_or("json");
-
-        // Repos that failed during this search. Reported to the caller: an
-        // agent that never sees the server log cannot otherwise distinguish a
-        // broken store from a repo that holds no match.
-        let mut literal_warnings: Vec<String> = Vec::new();
-
-        // Auto-regex promotion: detect code patterns that BM25 would destroy
-        let user_set_regex = request.regex.unwrap_or(false);
-        let user_set_phrase = request.phrase.unwrap_or(false);
-        let auto_promoted =
-            !user_set_regex && !user_set_phrase && looks_like_code_pattern(&request.query);
-
-        let (effective_query, effective_regex) = if auto_promoted {
-            let escaped = regex::escape(&request.query);
-            // Relax whitespace to \s+ so "foo = null" → "foo\s+=\s+null"
-            // regex::escape does not escape spaces, so replace literal spaces.
-            let relaxed = escaped.replace(' ', r"\s+");
-            (relaxed, true)
-        } else {
-            (request.query.clone(), user_set_regex)
-        };
-
-        tracing::debug!(
-            "MCP literal_search: query='{}', regex={:?}, phrase={:?}, limit={}, file_glob={:?}, language={:?}, format={}, multi={}",
-            request.query, request.regex, request.phrase, limit,
-            request.file_glob, request.language, output_format, ctx.is_multi
-        );
-
-        if ctx.needs_local_db {
-            if let Err(e) = self.ensure_database_exists() {
-                return Ok(CallToolResult::success(vec![Content::text(e)]));
-            }
-        }
-
-        // Pre-compute normalized project root for stripping absolute paths in glob matching
-        let lang_filter = request.language.clone();
-        let glob_filter = request.file_glob.clone();
-        let regex_enabled = effective_regex;
-        let snippet_regex = if regex_enabled {
-            Regex::new(&effective_query).ok()
-        } else {
-            None
-        };
-        let project_root_normalized = {
-            let root = crate::cache::normalize_path_str(self.project_path.to_str().unwrap_or(""));
-            root.trim_end_matches('/').to_string()
-        };
-
-        // Decide: BM25 path (for anchorable queries) or scan path (for tokenless regex
-        // or disjunctive OR patterns like TODO|FIXME|HACK that BM25 treats as AND).
-        let tokenless_regex = regex_enabled
-            && snippet_regex.is_some()
-            && (!regex_has_anchorable_token(&effective_query)
-                || regex_has_disjunctive_or(&effective_query));
-
-        let mut items: Vec<LiteralSearchResultItem> = if tokenless_regex {
-            // ── Scan path ──────────────────────────────────────────────
-            // Tokenless regex (e.g. \bfn\s+\w+) — BM25 cannot produce useful
-            // candidates. Scan all chunks sequentially, apply regex post-filter.
-            // Score is 0.0 for all results (no BM25 ranking applies).
-            tracing::debug!("literal_search: tokenless regex detected, using scan path");
-            if let Some(ref sv) = ctx.stores_vec {
-                // Multi-store scan
-                let mut items: Vec<LiteralSearchResultItem> = Vec::new();
-                for store_arc in sv {
-                    let store = store_arc.vector_store.read().await;
-                    let all_chunks = match store.iter_all_chunks() {
-                        Ok(chunks) => chunks,
-                        Err(_) => continue,
-                    };
-                    for (_, chunk) in all_chunks {
-                        if let Some(ref lang) = lang_filter {
-                            let file_lang = Language::from_path(std::path::Path::new(&chunk.path));
-                            if file_lang.name() != lang {
-                                continue;
-                            }
-                        }
-                        if let Some(ref glob) = glob_filter {
-                            let relative_path = chunk
-                                .path
-                                .strip_prefix(&project_root_normalized)
-                                .unwrap_or(&chunk.path)
-                                .trim_start_matches('/');
-                            if !simple_glob_match(glob, relative_path) {
-                                continue;
-                            }
-                        }
-                        if let Some((match_offset, snippet)) = match_line_for_literal(
-                            &chunk.content,
-                            &effective_query,
-                            snippet_regex.as_ref(),
-                        ) {
-                            let match_line = chunk.start_line + match_offset;
-                            items.push(LiteralSearchResultItem {
-                                path: chunk.path,
-                                start_line: match_line,
-                                end_line: match_line,
-                                snippet,
-                                score: 0.0, // No BM25 score — scan-path results are unranked
-                                kind: if chunk.kind.is_empty() {
-                                    None
-                                } else {
-                                    Some(chunk.kind)
-                                },
-                                signature: chunk.signature.filter(|s| !s.is_empty()),
-                            });
-                            if items.len() >= limit {
-                                break;
-                            }
-                        }
-                    }
-                    if items.len() >= limit {
-                        break;
-                    }
-                }
-                items
-            } else {
-                // Single-store scan
-                match self
-                    .with_vector_store_read_for(
-                        |store| {
-                            let all_chunks = store.iter_all_chunks()?;
-                            let mut items: Vec<LiteralSearchResultItem> = Vec::new();
-                            for (_, chunk) in all_chunks {
-                                if let Some(ref lang) = lang_filter {
-                                    let file_lang =
-                                        Language::from_path(std::path::Path::new(&chunk.path));
-                                    if file_lang.name() != lang {
-                                        continue;
-                                    }
-                                }
-                                if let Some(ref glob) = glob_filter {
-                                    let relative_path = chunk
-                                        .path
-                                        .strip_prefix(&project_root_normalized)
-                                        .unwrap_or(&chunk.path)
-                                        .trim_start_matches('/');
-                                    if !simple_glob_match(glob, relative_path) {
-                                        continue;
-                                    }
-                                }
-                                if let Some((match_offset, snippet)) = match_line_for_literal(
-                                    &chunk.content,
-                                    &effective_query,
-                                    snippet_regex.as_ref(),
-                                ) {
-                                    let match_line = chunk.start_line + match_offset;
-                                    items.push(LiteralSearchResultItem {
-                                        path: chunk.path,
-                                        start_line: match_line,
-                                        end_line: match_line,
-                                        snippet,
-                                        score: 0.0, // No BM25 score — scan-path results are unranked
-                                        kind: if chunk.kind.is_empty() {
-                                            None
-                                        } else {
-                                            Some(chunk.kind)
-                                        },
-                                        signature: chunk.signature.filter(|s| !s.is_empty()),
-                                    });
-                                    if items.len() >= limit {
-                                        break;
-                                    }
-                                }
-                            }
-                            Ok(items)
-                        },
-                        ctx.stores.clone(),
-                    )
-                    .await
-                {
-                    Ok(items) => items,
-                    Err(e) => {
-                        return Ok(CallToolResult::success(vec![Content::text(format!(
-                            "Error scanning chunks: {e:#}"
-                        ))]));
-                    }
-                }
-            }
-        } else {
-            // ── BM25 path ──────────────────────────────────────────────
-            // Note: regex=true uses BM25 for candidates, then post-filters with the
-            // actual regex on raw content (Tantivy's RegexQuery only works on individual
-            // tokens, not raw text — underscores/punctuation cause empty results).
-            //
-            // When regex is enabled, strip metacharacters from the BM25 query so
-            // Tantivy gets clean tokens (e.g. "class Cache" instead of "class \w+Cache\b").
-            let bm25_query = if regex_enabled {
-                let cleaned = extract_bm25_query_from_regex(&effective_query);
-                if cleaned.is_empty() {
-                    effective_query.clone()
-                } else {
-                    cleaned
-                }
-            } else {
-                effective_query.clone()
-            };
-            let fts_results = if let Some(ref sv) = ctx.stores_vec {
-                let sa = ctx.store_aliases.as_ref().unwrap();
-                let outcome = self
-                    .with_fts_store_read_multi(
-                        |fts_store| {
-                            if request.phrase.unwrap_or(false) {
-                                fts_store.search_phrase(&bm25_query, limit * 3)
-                            } else {
-                                fts_store.search(&bm25_query, limit * 3, None)
-                            }
-                        },
-                        sv.clone(),
-                        sa,
-                    )
-                    .await
-                    .unwrap_or_default();
-                for (alias, err) in &outcome.failures {
-                    let msg = format!("repo '{alias}' literal search failed: {err}");
-                    tracing::error!("MCP: {}", msg);
-                    literal_warnings.push(msg);
-                }
-                outcome.results
-            } else {
-                match self
-                    .with_fts_store_read_for(
-                        |fts_store| {
-                            if request.phrase.unwrap_or(false) {
-                                fts_store.search_phrase(&bm25_query, limit * 3)
-                            } else {
-                                fts_store.search(&bm25_query, limit * 3, None)
-                            }
-                        },
-                        ctx.stores.clone(),
-                    )
-                    .await
-                {
-                    Ok(r) => r,
-                    Err(e) => {
-                        return Ok(CallToolResult::success(vec![Content::text(format!(
-                            "Error searching: {e:#}"
-                        ))]));
-                    }
-                }
-            };
-
-            // Resolve chunk metadata and apply post-filters
-            if let Some(ref sv) = ctx.stores_vec {
-                // Multi-store: resolve chunks from all stores
-                let mut items: Vec<LiteralSearchResultItem> = Vec::new();
-                'outer: for fts_result in &fts_results {
-                    let sa = ctx.store_aliases.as_ref().unwrap();
-                    for (idx, store_arc) in sv.iter().enumerate() {
-                        let store = store_arc.vector_store.read().await;
-                        let looked_up = store.get_chunk(fts_result.chunk_id);
-                        if let Err(ref e) = looked_up {
-                            note_store_failure(&mut literal_warnings, sa, idx, "chunk lookup", e);
-                        }
-                        if let Some(chunk) = looked_up.ok().flatten() {
-                            if let Some(ref lang) = lang_filter {
-                                let file_lang =
-                                    Language::from_path(std::path::Path::new(&chunk.path));
-                                if file_lang.name() != lang {
-                                    continue;
-                                }
-                            }
-                            if let Some(ref glob) = glob_filter {
-                                let relative_path = chunk
-                                    .path
-                                    .strip_prefix(&project_root_normalized)
-                                    .unwrap_or(&chunk.path)
-                                    .trim_start_matches('/');
-                                if !simple_glob_match(glob, relative_path) {
-                                    continue;
-                                }
-                            }
-                            let match_info = match_line_for_literal(
-                                &chunk.content,
-                                &effective_query,
-                                snippet_regex.as_ref(),
-                            );
-                            if regex_enabled && match_info.is_none() {
-                                continue;
-                            }
-                            let (match_offset, snippet) = match_info.unwrap_or_else(|| {
-                                (0, chunk.content.lines().next().unwrap_or("").to_string())
-                            });
-                            let match_line = chunk.start_line + match_offset;
-                            items.push(LiteralSearchResultItem {
-                                path: chunk.path,
-                                start_line: match_line,
-                                end_line: match_line,
-                                snippet,
-                                score: fts_result.score,
-                                kind: if chunk.kind.is_empty() {
-                                    None
-                                } else {
-                                    Some(chunk.kind)
-                                },
-                                signature: chunk.signature.filter(|s| !s.is_empty()),
-                            });
-                            if items.len() >= limit {
-                                break 'outer;
-                            }
-                            break; // Found in this store
-                        }
-                    }
-                }
-                items
-            } else {
-                match self
-                    .with_vector_store_read_for(
-                        |store| {
-                            // Resolve chunk metadata first so a store `Err`
-                            // propagates to the error arm below ("Error
-                            // resolving search results") instead of silently
-                            // dropping the hit — `Ok(None)` alone is a true
-                            // miss ("chunk not in this store").
-                            let resolved: anyhow::Result<Vec<_>> = fts_results
-                                .iter()
-                                .map(|fts_result| {
-                                    let chunk = store.get_chunk(fts_result.chunk_id)?;
-                                    Ok((chunk, fts_result.score))
-                                })
-                                .collect();
-                            let items: Vec<LiteralSearchResultItem> = resolved?
-                                .into_iter()
-                                .filter_map(|(looked_up, score)| {
-                                    let chunk = looked_up?;
-                                    Some((chunk, score))
-                                })
-                                .filter(|(chunk, _)| {
-                                    if let Some(ref lang) = lang_filter {
-                                        let file_lang =
-                                            Language::from_path(std::path::Path::new(&chunk.path));
-                                        if file_lang.name() != lang {
-                                            return false;
-                                        }
-                                    }
-                                    if let Some(ref glob) = glob_filter {
-                                        let relative_path = chunk
-                                            .path
-                                            .strip_prefix(&project_root_normalized)
-                                            .unwrap_or(&chunk.path)
-                                            .trim_start_matches('/');
-                                        if !simple_glob_match(glob, relative_path) {
-                                            return false;
-                                        }
-                                    }
-                                    true
-                                })
-                                .take(limit)
-                                .filter_map(|(chunk, score)| {
-                                    let match_info = match_line_for_literal(
-                                        &chunk.content,
-                                        &effective_query,
-                                        snippet_regex.as_ref(),
-                                    );
-                                    if regex_enabled && match_info.is_none() {
-                                        return None;
-                                    }
-                                    let (match_offset, snippet) = match_info.unwrap_or_else(|| {
-                                        (0, chunk.content.lines().next().unwrap_or("").to_string())
-                                    });
-                                    let match_line = chunk.start_line + match_offset;
-                                    Some(LiteralSearchResultItem {
-                                        path: chunk.path,
-                                        start_line: match_line,
-                                        end_line: match_line,
-                                        snippet,
-                                        score,
-                                        kind: if chunk.kind.is_empty() {
-                                            None
-                                        } else {
-                                            Some(chunk.kind)
-                                        },
-                                        signature: chunk.signature.filter(|s| !s.is_empty()),
-                                    })
-                                })
-                                .collect();
-                            Ok(items)
-                        },
-                        ctx.stores.clone(),
-                    )
-                    .await
-                {
-                    Ok(items) => items,
-                    Err(e) => {
-                        return Ok(CallToolResult::success(vec![Content::text(format!(
-                            "Error resolving search results: {e:#}"
-                        ))]));
-                    }
-                }
-            }
-        };
-
-        // Prefix paths with alias for multi-repo identification
-        for item in &mut items {
-            item.path = ctx.prefix_result_path(&item.path);
-        }
-
-        // Compute low-confidence signal
-        let top_score = items.first().map(|i| i.score);
-        let (low_confidence, suggested_tool) =
-            compute_literal_low_confidence(top_score, &request.query);
-
-        // Build note
-        let note = if auto_promoted {
-            Some(format!(
-                "Query auto-promoted to regex mode (original: '{}', effective: '{}'). \
-                 The query contained code-like punctuation that BM25 would tokenize incorrectly.",
-                request.query, effective_query
-            ))
-        } else if low_confidence == Some(true) {
-            suggested_tool.as_ref().map(|tool| {
-                format!(
-                    "Top result has weak BM25 score; consider using `{}` for better matches.",
-                    tool
-                )
-            })
-        } else {
-            None
-        };
-
-        let response = LiteralSearchResponse {
-            results: items,
-            auto_promoted_to_regex: if auto_promoted { Some(true) } else { None },
-            note,
-            low_confidence,
-            suggested_tool: if low_confidence == Some(true) {
-                suggested_tool
-            } else {
-                None
-            },
-            warnings: if literal_warnings.is_empty() {
-                None
-            } else {
-                Some(literal_warnings)
-            },
-        };
-
-        // Instrument BM25 score for threshold calibration
-        if let Some(top) = response.results.first() {
-            tracing::debug!(
-                target: "codesearch::literal_confidence",
-                query = %request.query,
-                top_bm25_score = top.score,
-                result_count = response.results.len(),
-                "literal_search score sample"
-            );
-        }
-
-        // Format output
-        let output = if output_format == "grep" {
-            let mut lines: Vec<String> = Vec::new();
-            if response.auto_promoted_to_regex == Some(true) {
-                lines.push(
-                    "# auto-promoted to regex mode (query contained code-like punctuation)"
-                        .to_string(),
-                );
-            }
-            if response.low_confidence == Some(true) {
-                if let Some(ref hint) = response.suggested_tool {
-                    lines.push(format!("# low confidence — consider: {}", hint));
-                }
-            }
-            for item in &response.results {
-                lines.push(format!(
-                    "{}:{}:{}",
-                    item.path, item.start_line, item.snippet
-                ));
-            }
-            lines.join("\n")
-        } else {
-            serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string())
-        };
-
-        Ok(CallToolResult::success(vec![Content::text(output)]))
-    }
-
-    /// Internal implementation for index_status with optional project/group routing.
-    async fn index_status_impl(
-        &self,
-        project: Option<String>,
-        group: Option<String>,
-    ) -> Result<CallToolResult, McpError> {
-        // When no project/group specified in serve mode, return lightweight aggregated
-        // status WITHOUT opening any databases. Only a specific project/group request
-        // should trigger DB activation.
-        if project.is_none() && group.is_none() {
-            if let Some(ref serve_state) = self.serve_state {
-                let config = serve_state.config_snapshot();
-                let repo_count = config.repos.len();
-                // Count the virtual "all" group when repos are registered, so the
-                // summary doesn't read "0 group(s)" while `all` is actually available.
-                let group_count = config.groups.len() + if config.repos.is_empty() { 0 } else { 1 };
-                let statuses = serve_state.repo_statuses_lightweight();
-                let open_count = statuses
-                    .iter()
-                    .filter(|(_, r)| matches!(r.status, crate::serve::RepoStateLabel::Open))
-                    .count();
-                let warm_count = statuses
-                    .iter()
-                    .filter(|(_, r)| matches!(r.status, crate::serve::RepoStateLabel::Warm))
-                    .count();
-                let closed_count = statuses
-                    .iter()
-                    .filter(|(_, r)| matches!(r.status, crate::serve::RepoStateLabel::Closed))
-                    .count();
-
-                let status = if open_count + warm_count > 0 {
-                    "ready".to_string()
-                } else if repo_count > 0 {
-                    "idle".to_string()
-                } else {
-                    "no_repos".to_string()
-                };
-
-                let status_message = format!(
-                    "{} repo(s) registered, {} group(s). Open: {}, Warm: {}, Closed: {}.",
-                    repo_count, group_count, open_count, warm_count, closed_count
-                );
-
-                let response = IndexStatusResponse {
-                    indexed: open_count + warm_count > 0,
-                    status,
-                    status_message,
-                    total_chunks: 0, // Not available without opening DBs
-                    total_files: 0,
-                    model: self.model_type.short_name().to_string(),
-                    dimensions: 0,
-                    max_chunk_id: 0,
-                    db_path: format!("({} repos)", repo_count),
-                    project_path: format!("serve mode — {} repo(s)", repo_count),
-                    error_message: None,
-                    mode: self.mcp_mode(),
-                };
-
-                let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-                return Ok(CallToolResult::success(vec![Content::text(json)]));
-            }
-        }
-
-        // Resolve project/group routing — status is scope-free, allow unscoped fan-out
-        let ctx = match self.resolve_routing(&project, &group, true, "status").await {
-            Ok(c) => c,
-            Err(e) => return Ok(CallToolResult::success(vec![Content::text(e)])),
-        };
-
-        if ctx.needs_local_db {
-            let indexed = self.db_path.exists();
-
-            if !indexed {
-                let response = IndexStatusResponse {
-                    indexed: false,
-                    status: "not_indexed".to_string(),
-                    status_message: "No index found. Run 'codesearch index' or start with --create-index=true to automatically create one.".to_string(),
-                    total_chunks: 0,
-                    total_files: 0,
-                    model: "none".to_string(),
-                    dimensions: 0,
-                    max_chunk_id: 0,
-                    db_path: self.db_path.display().to_string(),
-                    project_path: self.project_path.display().to_string(),
-                    error_message: None,
-                    mode: self.mcp_mode(),
-                };
-                let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-                return Ok(CallToolResult::success(vec![Content::text(json)]));
-            }
-        }
-
-        if let Some(ref sv) = ctx.stores_vec {
-            // Multi-store: aggregate stats across all group members
-            let mut total_chunks = 0usize;
-            let mut total_files = 0usize;
-            let mut max_chunk_id = 0u32;
-            let mut dimensions = 0usize;
-            let mut all_indexed = true;
-            let aliases = ctx.aliases();
-            let mut stats_warnings: Vec<String> = Vec::new();
-            let mut failed_count = 0usize;
-
-            for (i, store_arc) in sv.iter().enumerate() {
-                let store = store_arc.vector_store.read().await;
-                match store.stats() {
-                    Ok(stats) => {
-                        total_chunks += stats.total_chunks;
-                        total_files += stats.total_files;
-                        if stats.max_chunk_id > max_chunk_id {
-                            max_chunk_id = stats.max_chunk_id;
-                        }
-                        if stats.dimensions > 0 {
-                            dimensions = stats.dimensions;
-                        }
-                        if !stats.indexed {
-                            all_indexed = false;
-                        }
-                    }
-                    // `all_indexed = false` alone renders identically to "still
-                    // warming" — the caller has no way to tell "wait" from "this
-                    // store is down". This is the tool whose job is reporting index
-                    // health, so it must not stay silent on the one signal that
-                    // matters here: bind the error, carry it, never `Err(_)`.
-                    Err(ref e) => {
-                        all_indexed = false;
-                        failed_count += 1;
-                        note_store_failure(&mut stats_warnings, aliases, i, "stats", e);
-                    }
-                }
-            }
-
-            let (status, status_message) =
-                index_status_summary(sv.len(), failed_count, total_chunks);
-
-            let response = IndexStatusResponse {
-                indexed: all_indexed,
-                status,
-                status_message,
-                total_chunks,
-                total_files,
-                model: self.model_type.short_name().to_string(),
-                dimensions,
-                max_chunk_id,
-                db_path: format!("({} repos)", sv.len()),
-                project_path: format!("group with {} repo(s)", sv.len()),
-                error_message: None,
-                mode: self.mcp_mode(),
-            };
-
-            return respond_with_object(&response, &stats_warnings);
-        }
-
-        // Single-store path
-        let stats = match self
-            .with_vector_store_read_for(
-                |store| store.stats().context("Error getting index stats"),
-                ctx.stores.clone(),
-            )
-            .await
-        {
-            Ok(s) => s,
-            Err(e) => {
-                let response = IndexStatusResponse {
-                    indexed: false,
-                    status: "error".to_string(),
-                    status_message: format!("{}", e),
-                    total_chunks: 0,
-                    total_files: 0,
-                    model: self.model_type.short_name().to_string(),
-                    dimensions: 0,
-                    max_chunk_id: 0,
-                    db_path: self.db_path.display().to_string(),
-                    project_path: self.project_path.display().to_string(),
-                    error_message: Some(format!("{}", e)),
-                    mode: self.mcp_mode(),
-                };
-                let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-                return Ok(CallToolResult::success(vec![Content::text(json)]));
-            }
-        };
-
-        // Determine status based on database state
-        let (status, status_message) = if stats.total_chunks == 0 {
-            (
-                "building".to_string(),
-                "Index is being built in the background. Searches may fail until indexing completes. Please check back in a few minutes.".to_string(),
-            )
-        } else {
-            (
-                "ready".to_string(),
-                "Index is ready for searching.".to_string(),
-            )
-        };
-
-        let response = IndexStatusResponse {
-            indexed: stats.indexed,
-            status,
-            status_message,
-            total_chunks: stats.total_chunks,
-            total_files: stats.total_files,
-            model: self.model_type.short_name().to_string(),
-            dimensions: stats.dimensions,
-            max_chunk_id: stats.max_chunk_id,
-            db_path: self.db_path.display().to_string(),
-            project_path: self.project_path.display().to_string(),
-            error_message: None,
-            mode: self.mcp_mode(),
-        };
-
-        let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-        Ok(CallToolResult::success(vec![Content::text(json)]))
-    }
-
-    /// List all registered projects and groups. Called by `status(kind="projects")`.
-    /// Build the `remote_projects` listing (opt-in mounts) for `list_projects`.
-    fn remote_projects_listing(
-        config: &crate::db_discovery::repos::ReposConfig,
-    ) -> Vec<RemoteProjectInfo> {
-        config
-            .mounted_remote_projects()
-            .into_iter()
-            .filter_map(|(name, target)| match target {
-                crate::db_discovery::repos::Target::RemoteProject {
-                    peer_name,
-                    peer,
-                    remote_alias,
-                } => Some(RemoteProjectInfo {
-                    name,
-                    peer: peer_name,
-                    remote_alias,
-                    peer_url: peer.url,
-                }),
-                _ => None,
-            })
-            .collect()
-    }
-
-    async fn list_projects(&self) -> Result<CallToolResult, McpError> {
-        let current_dir = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-
-        let serve_active = self.serve_state.is_some();
-        let serve_url = if serve_active {
-            Some(serve_url_from_env())
-        } else {
-            None
-        };
-
-        // When serve is active, use ServeState as source of truth for lock status
-        if let Some(ref serve_state) = self.serve_state {
-            let config = serve_state.config_snapshot();
-            let project_groups = config.project_groups();
-            let mut repos_info = Vec::new();
-            let mut list_warnings: Vec<String> = Vec::new();
-
-            for (alias, path) in &config.repos {
-                let db_path = path.join(crate::constants::DB_DIR_NAME);
-
-                let (total_chunks, total_files, model, lock_status, error) = if db_path.exists() {
-                    let (model_name, _dims) = read_model_metadata(&db_path);
-
-                    // For repos already opened in DashMap, use the live SharedStores for stats
-                    // WITHOUT opening a new VectorStore connection.
-                    // For unopened repos, just report metadata — do NOT open the DB.
-                    if let Some(stores) = serve_state.get_opened_stores(alias) {
-                        let stats_result = {
-                            let vs = stores.vector_store.read().await;
-                            vs.stats()
-                        };
-                        // `0 chunks` alone reads exactly like "not indexed yet" — the
-                        // repo may in fact be full and simply failing to answer (the
-                        // read-only-incident shape this branch exists for). Attribute
-                        // the failure to THIS repo rather than a top-level channel:
-                        // list_projects returns one entry per repo, so per-item is the
-                        // shape that actually matches the fan-out.
-                        // `repo_stats_from_result` carries only the part of this
-                        // decision that varies by Ok/Err — see its doc comment.
-                        // `record_stats_or_warn` wraps it so this call site cannot
-                        // silently drop the warning half without also breaking the
-                        // counts it returns — see its own doc comment.
-                        let (total_chunks, total_files, error) =
-                            record_stats_or_warn(stats_result, alias, &mut list_warnings);
-                        (
-                            total_chunks,
-                            total_files,
-                            model_name,
-                            serve_state
-                                .repo_lock_status(alias)
-                                .unwrap_or("unknown")
-                                .to_string(),
-                            error,
-                        )
-                    } else {
-                        // Repo NOT opened — read persisted stats from metadata.json
-                        let (md_chunks, md_files) = read_metadata_stats(&db_path);
-                        let lock_status = if crate::index::is_database_locked(&db_path) {
-                            "locked-externally".to_string()
-                        } else {
-                            "available".to_string()
-                        };
-                        (md_chunks, md_files, model_name, lock_status, None)
-                    }
-                } else {
-                    (0, 0, "not indexed".to_string(), "unknown".to_string(), None)
-                };
-
-                repos_info.push(RepoInfo {
-                    alias: alias.clone(),
-                    project_path: path.display().to_string(),
-                    database_path: db_path.display().to_string(),
-                    total_chunks,
-                    total_files,
-                    model,
-                    lock_status,
-                    groups: project_groups.get(alias).cloned().unwrap_or_default(),
-                    error,
-                });
-            }
-
-            let response = ListProjectsResponse {
-                repos: repos_info,
-                groups: config.groups_with_virtual_all(),
-                remote_projects: Self::remote_projects_listing(&config),
-                serve_active,
-                serve_url,
-                current_directory: current_dir.display().to_string(),
-            };
-
-            return respond_with_object(&response, &list_warnings);
-        }
-
-        // Stdio mode: fall back to disk-based lock detection
-        let config = load_repos_config().unwrap_or_default();
-        let project_groups = config.project_groups();
-        let mut repos_info = Vec::new();
-        for (alias, path) in &config.repos {
-            let db_path = path.join(crate::constants::DB_DIR_NAME);
-
-            // Get stats
-            let (total_chunks, total_files, model, lock_status) = if db_path.exists() {
-                let (model_name, dims) = read_model_metadata(&db_path);
-
-                let lock = if crate::index::is_database_locked(&db_path) {
-                    "conflicted"
-                } else {
-                    "available"
-                };
-
-                if let Ok(store) = VectorStore::new(&db_path, dims) {
-                    if let Ok(stats) = store.stats() {
-                        (
-                            stats.total_chunks,
-                            stats.total_files,
-                            model_name,
-                            lock.to_string(),
-                        )
-                    } else {
-                        (0, 0, model_name, lock.to_string())
-                    }
-                } else {
-                    (0, 0, model_name, "readonly".to_string())
-                }
-            } else {
-                (0, 0, "not indexed".to_string(), "unknown".to_string())
-            };
-
-            // Stdio mode is single-repo-at-a-time CLI usage, not the live multi-repo
-            // federation this fan-out fix targets — a stats() failure here is out of
-            // scope for this fix (VectorStore::new/stats failing locally is a different
-            // shape than a store going down mid-request in a shared serve process).
-            repos_info.push(RepoInfo {
-                alias: alias.clone(),
-                project_path: path.display().to_string(),
-                database_path: db_path.display().to_string(),
-                total_chunks,
-                total_files,
-                model,
-                lock_status,
-                groups: project_groups.get(alias).cloned().unwrap_or_default(),
-                error: None,
-            });
-        }
-
-        let response = ListProjectsResponse {
-            repos: repos_info,
-            groups: config.groups_with_virtual_all(),
-            remote_projects: Self::remote_projects_listing(&config),
-            serve_active,
-            serve_url,
-            current_directory: current_dir.display().to_string(),
-        };
-
-        let json = serde_json::to_string(&response).unwrap_or_else(|_| "{}".to_string());
-        Ok(CallToolResult::success(vec![Content::text(json)]))
-    }
 }
 
 // === Server Handler Implementation ===
@@ -5590,6 +1987,11 @@ impl CodesearchService {
         // CodesearchService (inside the find_impact.rs impl block), so the
         // per-module routers all resolve through the type, not the module.
         router.merge(Self::find_impact_router());
+        router.merge(Self::search_router());
+        router.merge(Self::find_router());
+        router.merge(Self::explore_router());
+        router.merge(Self::status_router());
+        router.merge(Self::get_chunk_router());
         router
     }
 }
